% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminance_and_contrast.R
\name{compute_apca_contrast}
\alias{compute_apca_contrast}
\title{Obliczanie kontrastu}
\usage{
compute_apca_contrast(fg, bg)
}
\arguments{
\item{fg}{wektor opisujący kolory pierwszoplanowe/tekstu
(przekazywany do \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}})}

\item{bg}{wektor opisujący kolory tła (przekazywany do \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}})}
}
\value{
wektor liczb z przedziału [0, 100]
}
\description{
Oblicza \emph{kontrast jasności} (\emph{Lightness Contrast}) zgodnie z definicją
stosowaną w APCA 0.1.9 4g (\emph{Accessible Perceptual Contrast Algorithm})
i WCAG 3.
}
\examples{
# przykłady z:
# <https://github.com/Myndex/SAPC-APCA/tree/master/documentation#testing-your-implementation>
compute_apca_contrast(
  c("#888", "#fff", "#000", "#aaa", "#123", "#def", "#123", "#444"),
  c("#fff", "#888", "#aaa", "#000", "#def", "#123", "#444", "#123"))
# test poprawności implementacji:
stopifnot(all((compute_apca_contrast(c("#888", "#fff",
                                       "#000", "#aaa",
                                       "#123", "#def",
                                       "#123", "#444"),
                                     c("#fff", "#888",
                                       "#aaa", "#000",
                                       "#def", "#123",
                                       "#444", "#123")) -
               c(63.056469930209424, -68.54146436644962,
                 58.146262578561334, -56.24113336839742,
                 91.66830811481631, -93.06770049484275,
                  8.32326136957393,  -7.526878460278154)) < 0.000000000001))
}
\seealso{
\url{https://github.com/Myndex/SAPC-APCA/tree/master/documentation},
\url{https://www.w3.org/WAI/GL/task-forces/silver/wiki/User:Myndex/APCA_model},
\code{\link[=compute_screen_luminance]{compute_screen_luminance()}}, ale też \code{\link[=compute_contrast_ratio]{compute_contrast_ratio()}}
}
