% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminance_and_contrast.R
\name{compute_contrast_ratio}
\alias{compute_contrast_ratio}
\title{Obliczanie kontrastu}
\usage{
compute_contrast_ratio(x, y)
}
\arguments{
\item{x}{wektor opisujący kolory (przekazywany do \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}})}

\item{y}{wektor opisujący kolory (przekazywany do \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}})}
}
\value{
wektor liczb z przedziału [1, 21]
}
\description{
Oblicza stosunek kontrastu (\emph{contrast ratio}) zgodnie z definicją stosowaną
we WCAG 2.x.
}
\examples{
compute_contrast_ratio(
  c("#888", "#000", "#123", "#123"),
  c("#fff", "#aaa", "#def", "#444"))
}
\seealso{
\url{https://www.w3.org/TR/UNDERSTANDING-WCAG20/visual-audio-contrast-contrast.html#contrast-ratiodef}
i \code{\link[=compute_relative_luminance]{compute_relative_luminance()}}, ale też \code{\link[=compute_apca_contrast]{compute_apca_contrast()}}
}
