% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminance_and_contrast.R
\name{compute_relative_luminance}
\alias{compute_relative_luminance}
\title{Obliczanie luminacji}
\usage{
compute_relative_luminance(x, errata = FALSE)
}
\arguments{
\item{x}{wektor opisujący kolory, przekazywany do \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}}

\item{errata}{opcjonalnie wartość logiczna (bez praktycznego znaczenia):
czy skorygować próg przejścia z funkcji wykładniczej na funkcję liniową tak,
aby był zgodny ze standardem IEC sRGB (tj. 0,04045 zamiast 0,03928)?}
}
\value{
wektor liczb z przedziału [0, 1]
}
\description{
Oblicza relatywną luminację koloru zgodnie z definicją stosowaną we WCAG 2.x.
}
\examples{
compute_relative_luminance(
  c("#888", "#fff", "#000", "#aaa", "#123", "#def", "#123", "#444"))
}
\seealso{
\url{https://www.w3.org/WAI/GL/wiki/Relative_luminance},
\url{https://www.color.org/srgb.pdf} oraz
\url{https://en.wikipedia.org/wiki/Relative_luminance}
i \code{\link[=compute_contrast_ratio]{compute_contrast_ratio()}}, ale też \code{\link[=compute_screen_luminance]{compute_screen_luminance()}}
}
