% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminance_and_contrast.R
\name{compute_screen_luminance}
\alias{compute_screen_luminance}
\title{Obliczanie luminacji}
\usage{
compute_screen_luminance(x)
}
\arguments{
\item{x}{wektor opisujący kolory, przekazywany do \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}}
}
\value{
wektor liczb z przedziału [0, 1]
}
\description{
Oblicza \emph{ekranową} luminację koloru zgodnie z definicją stosowaną w APCA
0.1.9 4g (\emph{Accessible Perceptual Contrast Algorithm}).
}
\examples{
compute_screen_luminance(
  c("#888", "#fff", "#000", "#aaa", "#123", "#def", "#123", "#444"))
}
\seealso{
\url{https://github.com/Myndex/SAPC-APCA/tree/master/documentation},
\url{https://www.w3.org/WAI/GL/task-forces/silver/wiki/User:Myndex/APCA_model},
\code{\link[=compute_apca_contrast]{compute_apca_contrast()}}, ale też \code{\link[=compute_relative_luminance]{compute_relative_luminance()}}
}
