% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette_fun.R
\name{create_palette_losy_fg_fun}
\alias{create_palette_losy_fg_fun}
\title{Skale kolorow}
\usage{
create_palette_losy_fg_fun(type, fgToChoose, wcag = "3")
}
\arguments{
\item{type}{ciąg znaków - typ skali}

\item{fgToChoose}{wektor kolorów (opisanych w dowolny sposób zrozumiały dla
\code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}), spośród których wybierane będą bardziej kontrastowe
w stosunku do kolorów skali, na którą wskazuje wartość argumentu \code{type}}

\item{wcag}{opcjonalnie wersja standardu WCAG (domyślnie 3), wg której ma być
obliczany kontrast (por. \link{compute_apca_contrast} i \link{compute_contrast_ratio})}
}
\value{
Funkcja zwracająca paletę kolorów (przyjmująca jako parametr liczbę
kolorów w zwracanej palecie, która jednak będzie działać tylko dla jednej
liczby kolorów - kompatybilnej z liczbą kolorów w skali kolorów podanego
typu, zwracanej przez \code{\link[=palette_losy]{palette_losy()}}).
}
\description{
Nieeksportowana fabryka funkcji do wykorzystania wewnętrznie przez
\code{\link[=scale_color_losy_fg]{scale_color_losy_fg()}} i \code{\link[=scale_fill_losy_fg]{scale_fill_losy_fg()}} w ramach mechanizmu skal kolorów
\emph{ggplot2}.
}
