% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneBezrobocie}
\alias{daneBezrobocie}
\title{Przykladowe dane}
\format{
Ramka danych z 21 wierszami i 2 kolumnami:
\itemize{
\item{\code{miesiac} - czynnik o dwudziestu jeden poziomach,}
\item{\code{bezrobocie} - odsetek absolwentów zarejestrowanych jako bezrobotni.}
}
}
\usage{
daneBezrobocie
}
\description{
Przykładowe zestawienie częstości bycia zarejestrowanym jako
bezrobotny ze względu na miesiąc od ukończenia szkoły.
}
\seealso{
\link{wykresBezrobocie}
}
\keyword{datasets}
