% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneBezrobocieOkresGrupaOdn}
\alias{daneBezrobocieOkresGrupaOdn}
\title{Przykladowe dane}
\format{
Ramka danych z 10 wierszami i 3 kolumnami:
\itemize{
\item{\code{grupa} - czynnik o wartościach \emph{Państwa szkoła} i \emph{Pozostałe [typ szkoły] z [JST stanowiąca grupę odniesienia]},}
\item{\code{okres_bezrobocia} - czynnik o poziomach \emph{Brak bezrobocia}, \emph{1 miesiąc}, \emph{2 miesiące}, \emph{3 miesiące}, \emph{4 miesiące},}
\item{\code{pct} - częstość danej długości okresu bezrobocia w ramach danej grupy.}
}
}
\usage{
daneBezrobocieOkresGrupaOdn
}
\description{
Przykładowe zestawienie częstości łącznej długości okresu bycia
zarejestrowanym jako bezrobotny (w okresie od września do grudnia roku
ukończenia szkoły) pomiędzy daną szkołą a odpowiadającą jej
\emph{grupą odniesienia} - do wykorzystania w raportach szkolnych.
}
\seealso{
\link{wykresBezrobocieOkresGrupaOdn}
}
\keyword{datasets}
