% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneBezrobociePlec}
\alias{daneBezrobociePlec}
\title{Przykladowe dane}
\format{
Ramka danych z 4 wierszami i 3 kolumnami:
\itemize{
\item{\code{plec} - czynnik o poziomach \emph{Mężczyźni}, \emph{Kobiety},}
\item{\code{bezrobocie} - czynnik o poziomach \emph{Zarejestrowany jako bezrobotny}, \emph{Brak statusu bezrobotnego},}
\item{\code{pct} - częstość danego statusu bezrobocia w ramach danej płci.}
}
}
\usage{
daneBezrobociePlec
}
\description{
Przykładowe zestawienie częstości bycia zarejestrowanym jako
bezrobotny w podziale ze względu na płeć.
}
\seealso{
\link{wykresBezrobociePlec}
}
\keyword{datasets}
