% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneBezrobocieZawod}
\alias{daneBezrobocieZawod}
\title{Przykladowe dane}
\format{
Ramka danych z 20 wierszami i 3 kolumnami:
\itemize{
\item{\code{nazwa_zaw} - czynnik z nazwami zawodów, kolejność poziomów którego odzwierciedla liczbę absolwentów w danym zawodzie,}
\item{\code{bezrobocie} - czynnik o poziomach \emph{Zarejestrowany jako bezrobotny}, \emph{Brak statusu bezrobotnego},}
\item{\code{pct} - częstość danego statusu bezrobocia w ramach danego zawodu.}
}
}
\usage{
daneBezrobocieZawod
}
\description{
Przykładowe zestawienie częstości bycia zarejestrowanym jako
bezrobotny w podziale ze względu na (najczęściej wybierane) zawody wyuczone.
}
\seealso{
\link{wykresBezrobocieZawod}
}
\keyword{datasets}
