% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneDyplomy}
\alias{daneDyplomy}
\title{Przykladowe dane}
\format{
Ramka danych z 12 wierszami i 3 kolumnami:
\itemize{
\item{\code{plec} - czynnik o poziomach \emph{Ogółem}, \emph{Mężczyźni}, \emph{Kobiety},}
\item{\code{dyplom} - czynnik o poziomach \emph{Świadectwo czeladnicze}, \emph{Dyplom zawodowy}, \emph{Tylko certyfikat kwalifikacji}, \emph{Brak certyfikatów i dyplomu}}
\item{\code{pct} - częstość uzyskiwanie dyplomów w ramach danej płci (lub ogółem).}
}
}
\usage{
daneDyplomy
}
\description{
Przykładowe zestawienie częstości uzyskiwania dyplomów
w podziale ze względu na płeć.
}
\seealso{
\link{wykresDyplomy}
}
\keyword{datasets}
