% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneDyplomyZawod}
\alias{daneDyplomyZawod}
\title{Przykladowe dane}
\format{
Ramka danych z 40 wierszami i 3 kolumnami:
\itemize{
\item{\code{nazwa_zaw} - czynnik z nazwami zawodów, kolejność poziomów którego odzwierciedla liczbę absolwentów w danym zawodzie,}
\item{\code{dyplom} - czynnik o poziomach \emph{Świadectwo czeladnicze}, \emph{Dyplom zawodowy}, \emph{Tylko certyfikat kwalifikacji}, \emph{Brak certyfikatów i dyplomu},}
\item{\code{pct} - częstość uzyskiwanie dyplomów w ramach danego zawodu.}
}
}
\usage{
daneDyplomyZawod
}
\description{
Przykładowe zestawienie częstości uzyskiwania dyplomów
w podziale ze względu na (najczęściej wybierane) zawody wyuczone.
}
\seealso{
\link{wykresDyplomyZawod}
}
\keyword{datasets}
