% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneKontMlodocPracNeduGrupaOdn}
\alias{daneKontMlodocPracNeduGrupaOdn}
\title{Przykladowe dane}
\format{
Ramka danych z 6 wierszami i 3 kolumnami:
\itemize{
\item{\code{grupa} - czynnik o wartościach \emph{Państwa szkoła} i \emph{Pozostałe [typ szkoły] z [JST stanowiąca grupę odniesienia]},}
\item{\code{kont_mlodoc_prac} - czynnik o wartościach \emph{Nieuczący się, kontynuujący pracę u danego praodawcy na podstawie umowy o pracę}, \emph{Nieuczący się, kontynuujący pracę u danego pracodawcy w formie innej niż umowa o pracę}, \emph{Nieuczący się, niekontynuujący pracy u danego pracodawcy},}
\item{\code{pct} - częstość danej sytuacji w ramach grupy.}
}
}
\usage{
daneKontMlodocPracNeduGrupaOdn
}
\description{
Przykładowe zestawienie częstości kontynuowania pracy u tego
samego pracodawcy, u którego absolwent był młodocianym pracownikiem, wśród
absolwentów nie kontynuujących nauki pomiędzy daną szkołą a odpowiadającą jej
\emph{grupą odniesienia} - do wykorzystania w raportach szkolnych.
}
\seealso{
\link{wykresKontMlodocPracNeduGrupaOdn}
}
\keyword{datasets}
