% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneKontynuacjeDyscyplinyPlec}
\alias{daneKontynuacjeDyscyplinyPlec}
\title{Przykladowe dane}
\format{
Ramka danych z 30 wierszami i 3 kolumnami:
\itemize{
\item{\code{plec} - czynnik o poziomach \emph{Ogółem}, \emph{Mężczyźni}, \emph{Kobiety},}
\item{\code{dyscyplina} - czynnik z nazwami dyscyplin, kolejność poziomów którego odzwierciedla częstość wyboru dyscyplin,}
\item{\code{pct} - częstość wyboru danej dziedziny w ramach danej płci (lub ogółem).}
}
}
\usage{
daneKontynuacjeDyscyplinyPlec
}
\description{
Przykładowe zestawienie częstości (najczęściej wybieranych)
dyscyplin kontynuowania nauki na studiach w podziale ze względu na płeć.
}
\seealso{
\link{wykresKontynuacjeDyscyplinyPlec}
}
\keyword{datasets}
