% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneKontynuacjeDziedziny}
\alias{daneKontynuacjeDziedziny}
\title{Przykladowe dane}
\format{
Ramka danych z 24 wierszami i 3 kolumnami:
\itemize{
\item{\code{plec} - czynnik o poziomach \emph{Ogółem}, \emph{Mężczyźni}, \emph{Kobiety},}
\item{\code{dziedzina} - czynnik o ośmiu poziomach,}
\item{\code{pct} - częstość wyboru danej dziedziny w ramach danej płci (lub ogółem).}
}
}
\usage{
daneKontynuacjeDziedziny
}
\description{
Przykładowe zestawienie częstości dziedzin kontynuowania nauki
na studiach w podziale ze względu na płeć.
}
\seealso{
\link{wykresKontynuacjeDziedziny}
}
\keyword{datasets}
