% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneKontynuacjeTypSzk}
\alias{daneKontynuacjeTypSzk}
\title{Przykladowe dane}
\format{
Ramka danych z 9 wierszami i 3 kolumnami:
\itemize{
\item{\code{plec} - czynnik o poziomach \emph{Ogółem}, \emph{Mężczyźni}, \emph{Kobiety},}
\item{\code{kontynuacja} - czynnik o poziomach \emph{w szkole policealnej}, \emph{w ramach KKZ}, \emph{na studiach},}
\item{\code{pct} - częstość wyboru danego typu kształcenia w ramach danej płci (lub ogółem).}
}
}
\usage{
daneKontynuacjeTypSzk
}
\description{
Przykładowe zestawienie częstości typu kontynuowania nauki
w podziale ze względu na płeć.
}
\seealso{
[wykresKontynuacjeTypSzk
}
\keyword{datasets}
