% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneKontynuacjeTypSzkPlec}
\alias{daneKontynuacjeTypSzkPlec}
\title{Przykladowe dane}
\format{
Ramka danych z 9 wierszami i 3 kolumnami:
\itemize{
\item{\code{plec} - czynnik o poziomach \emph{Ogółem}, \emph{Mężczyźni}, \emph{Kobiety},}
\item{\code{kontynuacja} - czynnik o poziomach \emph{W szkole policealnej}, \emph{W ramach KKZ}, \emph{Na studiach} (w zależności od typu szkoły może to też być: \emph{W ramach KKZ}, \emph{W liceum dla dorosłych}, \emph{W BS II})}
\item{\code{pct} - częstość wyboru danego typu kształcenia w ramach danej płci (lub ogółem).}
}
}
\usage{
daneKontynuacjeTypSzkPlec
}
\description{
Przykładowe zestawienie częstości typu kontynuowania nauki
w podziale ze względu na płeć.
}
\seealso{
\link{wykresKontynuacjeTypSzkPlec}
}
\keyword{datasets}
