% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneKontynuacjeTypSzkZawod}
\alias{daneKontynuacjeTypSzkZawod}
\title{Przykladowe dane}
\format{
Ramka danych z 30 wierszami i 3 kolumnami:
\itemize{
\item{\code{nazwa_zaw} - czynnik z nazwami zawodów, kolejność poziomów którego odzwierciedla liczbę absolwentów w danym zawodzie,}
\item{\code{kontynuacja} - czynnik o poziomach \emph{W szkole policealnej}, \emph{W ramach KKZ}, \emph{Na studiach},}
\item{\code{pct} - częstość wyboru danego typu kształcenia w ramach danego zawodu.}
}
}
\usage{
daneKontynuacjeTypSzkZawod
}
\description{
Przykładowe zestawienie częstości typu kontynuowania nauki
w podziale ze względu na (najczęściej wybierane) zawody wyuczone.
}
\seealso{
\link{wykresKontynuacjeTypSzkZawod}
}
\keyword{datasets}
