% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneMatury}
\alias{daneMatury}
\title{Przykladowe dane}
\format{
Ramka danych z 20 wierszami i 3 kolumnami:
\itemize{
\item{\code{zawod} - czynnik o dziesięciu poziomach,}
\item{\code{kontynuacja} - czynnik o poziomach \emph{Uzyskanie świadectwa dojrzałości}, \emph{Brak świadectwa dojrzałości},}
\item{\code{pct} - częstość danego wyniku matury w ramach danego zawodu.}
}
}
\usage{
daneMatury
}
\description{
Przykładowe zestawienie wyników matury w podziale ze względu na
wyuczony zawód.
}
\seealso{
\link{wykresMatury}
}
\keyword{datasets}
