% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneMaturyPlec}
\alias{daneMaturyPlec}
\title{Przykladowe dane}
\format{
Ramka danych z 6 wierszami i 3 kolumnami:
\itemize{
\item{\code{plec} - czynnik o poziomach \emph{Ogółem}, \emph{Mężczyźni}, \emph{Kobiety},}
\item{\code{dyplom} - czynnik o poziomach \emph{Uzyskanie świadectwa dojrzałości}, \emph{Brak świadectwa dojrzałości},}
\item{\code{pct} - częstość danego wyniku matury w ramach danej płci (lub ogółem).}
}
}
\usage{
daneMaturyPlec
}
\description{
Przykładowe zestawienie wyników matury w podziale ze względu na
płeć.
}
\seealso{
\link{wykresMaturyPlec}
}
\keyword{datasets}
