% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneMaturyZawod}
\alias{daneMaturyZawod}
\title{Przykladowe dane}
\format{
Ramka danych z 20 wierszami i 3 kolumnami:
\itemize{
\item{\code{nazwa_zaw} - czynnik z nazwami zawodów, kolejność poziomów którego odzwierciedla liczbę absolwentów w danym zawodzie,}
\item{\code{dyplom} - czynnik o poziomach \emph{Uzyskanie świadectwa dojrzałości}, \emph{Brak świadectwa dojrzałości},}
\item{\code{pct} - częstość danego wyniku matury w ramach danego zawodu.}
}
}
\usage{
daneMaturyZawod
}
\description{
Przykładowe zestawienie wyników matury w podziale ze względu na
(najczęściej wybierane) zawody wyuczone.
}
\seealso{
\link{wykresMaturyZawod}
}
\keyword{datasets}
