% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneMigracjeWoj}
\alias{daneMigracjeWoj}
\title{Przykladowe dane}
\format{
Ramka danych z 17 wierszami i 2 kolumnami:
\itemize{
\item{\code{teryt_woj} - kody (1/2-cyforwe) TERYT województw,}
\item{\code{pct} - częstość zamieszkiwania absolwentów na terenie danego województwa (na podstawie danych z ZUS).}
}
}
\usage{
daneMigracjeWoj
}
\description{
Przykładowe zestawienie rozkładu województw zamieszkania
absolwentów (z pewnego województwa) na podstawie danych z ZUS.
}
\seealso{
\link{wykresMigracjeWoj}, \link{sfWoj}
}
\keyword{datasets}
