% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{danePlec}
\alias{danePlec}
\title{Przykladowe dane}
\format{
Ramka danych z 14 wierszami i 3 kolumnami:
\itemize{
\item{\code{typ_szk} - czynnik o siedmiu poziomach,}
\item{\code{plec} - czynnik o poziomach \emph{Mężczyźni}, \emph{Kobiety},}
\item{\code{pct} - częstość danej płci w ramach danego typu szkoły.}
}
}
\usage{
danePlec
}
\description{
Przykładowe zestawienie płci w podziale ze względu na typ
ukończonej szkoły.
}
\seealso{
\link{wykresPlec}
}
\keyword{datasets}
