% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{danePracaOkresGrupaOdn}
\alias{danePracaOkresGrupaOdn}
\title{Przykladowe dane}
\format{
Ramka danych z 6 wierszami i 3 kolumnami:
\itemize{
\item{\code{grupa} - czynnik o wartościach \emph{Państwa szkoła} i \emph{Pozostałe [typ szkoły] z [JST stanowiąca grupę odniesienia]},}
\item{\code{okres_pracy} - czynnik o wartościach \emph{Brak pracy}, \emph{Praca przez część okresu}, \emph{Praca przez cały okres},}
\item{\code{pct} - częstość danej długości okresu w ramach grupy.}
}
}
\usage{
danePracaOkresGrupaOdn
}
\description{
Przykładowe zestawienie części okresu posiadania pracy
(w ramach okresu od września do grudnia roku ukończenia szkoły) pomiędzy daną
szkołą a odpowiadającą jej \emph{grupą odniesienia} - do wykorzystania w raportach
szkolnych.
}
\seealso{
\link{wykresPracaOkresGrupaOdn}
}
\keyword{datasets}
