% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{danePrzeplywyStatusy}
\alias{danePrzeplywyStatusy}
\title{Przykladowe dane}
\format{
Ramka danych z 375 wierszami i 5 kolumnami:
\itemize{
\item{\code{idKombinacji} - unikalny identyfikator ścieżki (kombinacji statusów),}
\item{\code{okres} - czynnik o wartościach \emph{12.2019}, \emph{12.2021}, \emph{12.2023}, identyfikujący punkt czasu,}
\item{\code{status} - czynnik o poziomach \emph{Tylko nauka}, \emph{Nauka i praca}, \emph{Tylko praca}, \emph{Bezrobocie}, \emph{Brak danych o aktywności},}
\item{\code{pct} - częstość występowania danej ścieżki (kombinacji statusów),}
\item{\code{pctSstatus} -  częstość występowania danego statusu w danym punkcie czasu.}
}
}
\usage{
danePrzeplywyStatusy
}
\description{
Przykładowe zestawienie częstości występowania poszczególnych
ścieżek edukacyjno-zawodowych, definiowanych jako kombinacje statusu
edukacyjno-zawodowego w trzech punktach czasu: w 6, 30 i 54 miesiące od
ukończenia szkoły.
}
\seealso{
\link{wykresPrzeplywyStatusy}
}
\keyword{datasets}
