% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{danePrzychodyGrupaOdn}
\alias{danePrzychodyGrupaOdn}
\title{Przykladowe dane}
\format{
Ramka danych z 2 wierszami i 6 kolumnami:
\itemize{
\item{\code{grupa} - czynnik o wartościach \emph{Państwa szkoła} i \emph{Pozostałe [typ szkoły] z [JST stanowiąca grupę odniesienia]},}
\item{\code{c05} - 5. centyl rozkładu przychodów w ramach grupy,}
\item{\code{c25} - 25. centyl (1. kwartyl) rozkładu przychodów w ramach grupy,}
\item{\code{c50} - mediana rozkładu przychodów w ramach grupy,}
\item{\code{c75} - 75. centyl (3. kwartyl) rozkładu przychodów w ramach grupy,}
\item{\code{c95} - 95. centyl rozkładu przychodów w ramach grupy.}
}
}
\usage{
danePrzychodyGrupaOdn
}
\description{
Przykładowe zestawienie parametrów rozkładu przychodów
z pracy (etatowej) pomiędzy daną szkołą a odpowiadającą jej
\emph{grupą odniesienia} - do wykorzystania w raportach szkolnych.
}
\seealso{
\link{wykresPrzychodyGrupaOdn}
}
\keyword{datasets}
