% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneStatusy}
\alias{daneStatusy}
\title{Przykladowe dane}
\format{
Ramka danych ze 105 wierszami i 3 kolumnami:
\itemize{
\item{\code{miesiac} - czynnik o dwudziestu jeden poziomach,}
\item{\code{status} - czynnik o poziomach \emph{Tylko nauka}, \emph{Nauka i praca}, \emph{Tylko praca}, \emph{Bezrobocie}, \emph{Brak danych o aktywności},}
\item{\code{pct} - częstość danego statusu w ramach danego miesiąca od ukończenia szkoły.}
}
}
\usage{
daneStatusy
}
\description{
Przykładowe zestawienie statusów edukacyjno-zawodowych ze
względu na miesiąc od ukończenia szkoły.
}
\seealso{
\link{wykresStatusy}
}
\keyword{datasets}
