% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneStatusyGrupaOdn}
\alias{daneStatusyGrupaOdn}
\title{Przykladowe dane}
\format{
Ramka danych z 10 wierszami i 3 kolumnami:
\itemize{
\item{\code{grupa} - czynnik o wartościach \emph{Państwa szkoła} i \emph{Pozostałe [typ szkoły] z [JST stanowiąca grupę odniesienia]},}
\item{\code{status} - czynnik o poziomach \emph{Tylko nauka}, \emph{Nauka i praca}, \emph{Tylko praca}, \emph{Bezrobocie}, \emph{Brak danych o aktywności},}
\item{\code{pct} - częstość danego statusu w ramach danej grupy.}
}
}
\usage{
daneStatusyGrupaOdn
}
\description{
Przykładowe zestawienie statusów edukacyjno-zawodowych pomiędzy
daną szkołą (ew. danym zawodem w danej szkole) a odpowiadającą jej
\emph{grupą odniesienia} - do wykorzystania w raportach szkolnych.
}
\seealso{
\link{wykresStatusyGrupaOdn}
}
\keyword{datasets}
