% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneStatusyPlec}
\alias{daneStatusyPlec}
\title{Przykladowe dane}
\format{
Ramka danych z 10 wierszami i 3 kolumnami:
\itemize{
\item{\code{plec} - czynnik o poziomach \emph{Mężczyźni}, \emph{Kobiety},}
\item{\code{status} - czynnik o poziomach \emph{Tylko nauka}, \emph{Nauka i praca}, \emph{Tylko praca}, \emph{Bezrobocie}, \emph{Brak danych o aktywności},}
\item{\code{pct} - częstość danego statusu w ramach danej płci.}
}
}
\usage{
daneStatusyPlec
}
\description{
Przykładowe zestawienie statusów edukacyjno-zawodowych ze
względu na płeć.
}
\seealso{
\link{wykresStatusyPlec}
}
\keyword{datasets}
