% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneStatusyZawod}
\alias{daneStatusyZawod}
\title{Przykladowe dane}
\format{
Ramka danych z 50 wierszami i 3 kolumnami:
\itemize{
\item{\code{nazwa_zaw} - czynnik z nazwami zawodów, kolejność poziomów którego odzwierciedla liczbę absolwentów w danym zawodzie,}
\item{\code{status} - czynnik o poziomach \emph{Tylko nauka}, \emph{Nauka i praca}, \emph{Tylko praca}, \emph{Bezrobocie}, \emph{Brak danych o aktywności},}
\item{\code{pct} - częstość danego statusu w ramach danego zawodu.}
}
}
\usage{
daneStatusyZawod
}
\description{
Przykładowe zestawienie statusów edukacyjno-zawodowych ze
względu na (najczęściej wybierane) zawody wyuoczne.
}
\seealso{
\link{wykresStatusyZawod}
}
\keyword{datasets}
