% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneTypySzkol}
\alias{daneTypySzkol}
\title{Przykladowe dane}
\format{
Ramka danych z 7 wierszami i 3 kolumnami:
\itemize{
\item{\code{rok_abs} - rok ukończenia szkoły (liczba),}
\item{\code{typ_szk} - czynnik o siedmiu poziomach,}
\item{\code{pct} - częstość danego typu szkoły.}
}
}
\usage{
daneTypySzkol
}
\description{
Przykładowe zestawienie typów ukończonych szkół.
}
\seealso{
\link{wykresTypySzkol}
}
\keyword{datasets}
