% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_zmien_rok_w_podpisie.R
\name{dodaj_rok_w_podpisie}
\alias{dodaj_rok_w_podpisie}
\alias{zmien_rok_w_podpisie}
\title{Dodawanie podpisu do wykresu}
\usage{
dodaj_rok_w_podpisie(rok)

zmien_rok_w_podpisie(g, rok)
}
\arguments{
\item{rok}{wartość, która ma zostać podstawiona jako rok pozyskania danych}

\item{g}{wykres \emph{\link[ggplot2:ggplot]{ggplot2}}}
}
\value{
\code{dodaj_rok_w_podpisie()} - listę klasy \emph{labels};
\code{zmien_rok_w_podpisie} - wykres \emph{\link[ggplot2:ggplot]{ggplot2}}
}
\description{
Dodawanie do wykresu lub zmienianie już istniejącego podpisu (\emph{caption})
zawierającego formułkę dot. źródła danych.

\code{dodaj_rok_w_podpisie()} - do wykorzystania w ramach procesu tworzenia
wykresu \emph{\link[ggplot2:ggplot]{ggplot2}} przy pomocy operatora \code{+},

\code{zmien_rok_w_podpisie()} - do zmodyfikowania już utworzonego (szablonu)
wykresu \emph{\link[ggplot2:ggplot]{ggplot2}} (np. w ramach pipe'ów)
}
\examples{
library(ggplot2)
p <- ggplot(daneTypySzkol,
            aes(pct, factor(rok_abs), fill = typ_szk,
                label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE))
p # wykres bez dodanej formułki

# dodawanie formułki:
p + dodaj_rok_w_podpisie(1234)
# ale
zmien_rok_w_podpisie(p, 1234)
# lub
p |> zmien_rok_w_podpisie(1234)
}
