% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_pl.R
\name{label_percent_pl}
\alias{label_percent_pl}
\alias{percent_pl}
\title{Formatowanie procentow}
\usage{
label_percent_pl(digits = NULL, hideBelow = 0, ...)

percent_pl(x, digits = NULL, hideBelow = 0, ...)
}
\arguments{
\item{digits}{liczba miejsc po przecinku, do której mają zostać zaokrąglone
wartości procentowe}

\item{hideBelow}{opcjonalnie wartość liczbowa: wartości poniżej tego progu
będą przez zwracaną funkcję \emph{ukrywane} (zamieniane na pusty ciąg znaków)}

\item{...}{argumenty przekazywane dalej do \code{\link[scales:label_percent]{scales::label_percent()}}
(z wyjątkiem ew. argumentów \code{decimal.mark} i \code{big.mark}, których wartość
jest ustawiana na sztywno odpowiednio na ',' i ' ')}

\item{x}{wektor liczb, które mają zostać sformatowane}
}
\value{
funkcja, p. \code{\link[scales:label_percent]{scales::label_percent()}}
}
\description{
\code{label_percent_pl()} wywołuje \code{\link[scales:label_percent]{scales::label_percent()}} z zestawem argumentów
odpowiadających polskiej ortografii, dając też bardziej typowy sposób
wskazywania pożądanego zaokrąglenia i możliwość \emph{ukrywania} wartości poniżej
zadanego progu;  jest przeznaczona przede wszystkim do używania w ramach
wywołań funkcji \code{scale_(...)()}, poprzez przekazanie
\strong{wyniku jej wywołania} do argumentu \code{labels}.

\code{percent_pl()} stanowi wygodny sposób na uniknięcie konieczności stosowania
dziwacznej na pierwszy rzut oka składni \code{label_percent_pl()(wektor_liczb)},
dzięki czemu można jej wygodnie używać np. definiując przypisanie wartości
do elementu \code{label} wykresu w ramach wywołania \code{\link[ggplot2:aes]{ggplot2::aes()}}.
}
\examples{
pct <- c(0, seq(0.015, 0.045, by = 0.01), seq(0.2, 1, by = 0.2))

percent_pl(pct)
percent_pl(pct, digits = 0)
percent_pl(pct, digits = 0, hideBelow = 0.05)

# dla porównania:
scales::percent(pct)
scales::percent(pct, decimal.mark = ",")
scales::percent(pct, accuracy = 1)

# label_percent_pl()
label_percent_pl()(pct)
label_percent_pl(digits = 0, hideBelow = 0.05)(pct)
# typowo do użycia przy formatowaniu etykiet osi wykresu (tu: poziomej):
library(ggplot2)
ggplot(daneTypySzkol,
       aes(pct, factor(rok_abs), fill = typ_szk)) +
  geom_col(position = position_stack(reverse = TRUE)) +
  scale_x_continuous(NULL, breaks = seq(0, 1, length.out = 9),
                     labels = label_percent_pl(digits = 1))
}
\seealso{
\code{\link[=zmien_prog_pokazywania_etykiet]{zmien_prog_pokazywania_etykiet()}},
\link{wykresBezrobocie}, \link{wykresBezrobocieOkresGrupaOdn}, \link{wykresBezrobociePlec},
\link{wykresBezrobocieZawod},
\link{wykresDyplomyPlec}, \link{wykresDyplomyZawod},
\link{wykresFormyZatrudnGrupaOdn}, \link{wykresKontMlodocPracEduGrupaOdn},
\link{wykresKontMlodocPracNeduGrupaOdn},
\link{wykresKontynuacjeDziedzinyPlec}, \link{wykresKontynuacjeDyscyplinyPlec},
\link{wykresKontynuacjeTypSzk}, \link{wykresKontynuacjeTypSzkPlec},
\link{wykresKontynuacjeTypSzkZawod},
\link{wykresMaturyPlec}, \link{wykresMaturyZawod},
\link{wykresPlec}, \link{wykresPracaOkresGrupaOdn}, \link{wykresPrzychodyGrupaOdn},
\link{wykresStatusy}, \link{wykresStatusyPlec}, \link{wykresStatusyZawod},
\link{wykresStatusyGrupaOdn}, \link{wykresTypySzkol}
}
