% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette_fun.R
\name{pal_seq_gradient_fg}
\alias{pal_seq_gradient_fg}
\title{Skale kolorow}
\usage{
pal_seq_gradient_fg(low, high, fgToChoose, wcag = "3")
}
\arguments{
\item{low, high}{Colours for low and high ends of the gradient.}

\item{fgToChoose}{wektor kolorów (opisanych w dowolny sposób zrozumiały dla
\code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}), spośród których wybierane będą bardziej kontrastowe
w stosunku do kolorów skali}

\item{wcag}{opcjonalnie wersja standardu WCAG (domyślnie 3), wg której ma być
obliczany kontrast (por. \link{compute_apca_contrast} i \link{compute_contrast_ratio})}
}
\value{
\emph{Ciągła} paleta kolorów \emph{ggplot2}.
}
\description{
Nieeksportowana fabryka funkcji do wykorzystania wewnętrznie przez
\code{\link[=scale_color_gradient_fg]{scale_color_gradient_fg()}} w ramach mechanizmu skal kolorów \emph{ggplot2}.
}
