% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_losy.R
\name{palette_losy}
\alias{palette_losy}
\title{Skale kolorow}
\usage{
palette_losy(
  type = c("statusy", "plec", "matura", "dyplomyZaw", "typySzkol", "kontynuacje")
)
}
\arguments{
\item{type}{ciąg znaków - typ skali}
}
\description{
Zwraca skale kolorów wykorzystywane w monitoringu karier absolwentów szkół
ponadpodstawowych.#'
}
