% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminance_and_contrast.R
\name{pick_more_contrasting}
\alias{pick_more_contrasting}
\title{Obliczanie kontrastu}
\usage{
pick_more_contrasting(bg, fg = c("#000000", "#FFFFFF"), wcag = c("3", "2"))
}
\arguments{
\item{bg}{wektor opisujący kolory tła (przekazywany do \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}})}

\item{fg}{opcjonalnie wektor opisujący kolory pierwszoplanowe/tekstu,
(przekazywany do \code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}) spośród których nastąpi wybór
(domyślnie biały lub czarny)}

\item{wcag}{opcjonalnie wersja standardu WCAG (domyślnie 3), wg której ma być
obliczany kontrast (por. \link{compute_apca_contrast} i \link{compute_contrast_ratio})}
}
\value{
wektor tekstowy z kodami RGB wybranych kolorów
pierwszoplanowych/tekstu
}
\description{
Dla zadanego wektora kolorów tła wybiera bardziej kontrastowy z zadanego
zestawu możliwych do wyboru kolorów pierwszoplanowych/tekstu.
}
\examples{
# różnice między WCAG 2 a WCAG 3 są znaczne:
pick_more_contrasting(palette_losy("statusy"), wcag = "3")
pick_more_contrasting(palette_losy("statusy"), wcag = "2")
}
