% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podmien_dane_wykresu.R
\name{podmien_dane_wykresu}
\alias{podmien_dane_wykresu}
\title{Podmiana danych wykresu}
\usage{
podmien_dane_wykresu(g, dane, jesliRozne = c("NULL", "porownanie", "stop"))
}
\arguments{
\item{g}{wykres \emph{\link[ggplot2:ggplot]{ggplot2}}}

\item{dane}{dane - jeśli operacja podmiany ma się udać, musi to być ramka
danych o strukturze takiej samej, jak dane już zawarte w wykresie przekazanym
argumentem \code{g}}

\item{jesliRozne}{ciąg znaków (domyślnie "NULL") wskazujący, jak ma się
zachować funkcja, jeśli obiekt przekazany argumentem \code{dane} ma strukturę nie
pasującą do wymagań wykresu przekazanego argumentem \code{g}: "NULL" - zwrócić
\code{NULL}, "porownanie" - zwrócić wynik porównania obiektu przekazanego
argumentem \code{dane} z danymi zawartymi w wykresie, "stop" - wygenerować błąd}
}
\value{
obiekt klasy \link[ggplot2:ggplot]{ggplot2::ggplot} lub \code{NULL} lub wektor ciągów znaków
}
\description{
Funkcja pozwala podjąć próbę podmiany danych w wykresie \emph{ggplot2}, obsługując
przy tym sytuację, gdy struktura danych nie pasuje do wymagań wykresu.
}
\details{
Do porównania ze sobą struktury danych w wykresie i podmienianych danych
używana jest funkcja \code{\link[=porownaj_strukture_danych]{porownaj_strukture_danych()}}.

Należy zwrócić uwagę, że ponieważ funkcja \code{podmien_dane_wykresu()} (wywołana
z domyślną wartością argumentu \code{jesliRozne}) może zwrócić wykres
\emph{\link[ggplot2:ggplot]{ggplot2}}, ale może też zwrócić \code{NULL}, więc wszelkie ew.
zmiany formatu wykresu dokonywane poprzez wywołania funkcji (np.
\code{\link[=zmien_wielkosc_czcionek]{zmien_wielkosc_czcionek()}}, \code{\link[=zmien_prog_pokazywania_etykiet]{zmien_prog_pokazywania_etykiet()}},
\code{\link[=zmien_rok_w_podpisie]{zmien_rok_w_podpisie()}}, czy \code{\link[=theme_losy]{theme_losy()}} - tej ostatniej w składni z \code{+})
należy dokonać przed jej wywołaniem (por. przykłady użycia poniżej).
}
\examples{
library(ggplot2)
podmien_dane_wykresu(wykresStatusy, daneStatusy)
# poniższe wywołanie zwróci NULL
podmien_dane_wykresu(wykresStatusy, daneStatusy[, -1])
# jeśli wykres ma być w jakiś sposób modyfikowany, to należy to zrobić
# przed wywołaniem podmien_dane_wykresu():
wykresStatusy |>
  zmien_wielkosc_czcionek(baseSize = 14) |>
  zmien_prog_pokazywania_etykiet(hideBelow = 0.04) |>
  zmien_rok_w_podpisie(2025) |>
  podmien_dane_wykresu(daneStatusy[, -1])
# można też się dowiedzieć, na czym polega rozbieżność struktury danych:
podmien_dane_wykresu(wykresStatusy, daneStatusy[, -1], jesliRozne = "porownanie")
}
\seealso{
\code{\link[=porownaj_strukture_danych]{porownaj_strukture_danych()}}
}
