% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porownaj_strukture_danych.R
\name{porownaj_strukture_danych}
\alias{porownaj_strukture_danych}
\title{Podmiana danych wykresu}
\usage{
porownaj_strukture_danych(x, y)
}
\arguments{
\item{x}{wzorcowy obiekt}

\item{y}{obiekt do porównania z \code{x}}
}
\value{
\code{TRUE} lub wektor tekstowy z opisem różnic (por. \code{\link[=all.equal]{all.equal()}})
}
\description{
Funkcja porównuje ze sobą strukturę dwóch obiektów (typowo ramek danych)
i raportuje ewentualne rozbieżności (jest przy tym znacznie mniej
restrykcyjna w porównaniach, niż \code{\link[=all.equal]{all.equal()}}).
}
\details{
Funkcja przeprowadza tylko cztery testy zgodności, adekwatne w kontekście
ustalania, czy na podstawie danych z obiektu \code{y} da się narysować wykres,
który został utworzony na podstawie danych z obiektu \code{x}:
\itemize{
\item Jeśli \code{x} jest ramką danych (lub obiektem \emph{ramkodanowym}, jak np.
\emph{tibble}), to czy \code{y} również jest ramką danych (obiektem
\emph{ramkodanowym}),
\item Czy wszystkie elementy (identyfikowane po ich nazwach - zwykle kolumny)
z \code{x} występują w \code{y},
\item Czy wszystkie elementy (identyfikowane po ich nazwach - zwykle kolumny),
które w \code{x} są wektorami liczbowymi lub logicznymi są również wektorami
liczbowymi lub logicznymi w \code{y},
\item Czy wszystkie elementy (identyfikowane po ich nazwach - zwykle kolumny),
które w \code{x} są wektorami tekstowymi lub czynnikami są również wektorami
liczbowymi lub czynnikami w \code{y}.
}
}
\examples{
df1 <- data.frame(a = 1:5, b = letters[1:5], c = factor(1:5),
                  d = sample(c(FALSE, TRUE), 5, replace = TRUE),
                  e = 1L:5L)
df2 <- data.frame(b = 1:5, a = letters[1:5], d = factor(1:5),
                  c = sample(c(FALSE, TRUE), 5, replace = TRUE))
porownaj_strukture_danych(df1, df1)
porownaj_strukture_danych(df1, df2)
}
\seealso{
\code{\link[=podmien_dane_wykresu]{podmien_dane_wykresu()}}
}
