% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_fill_losy_fg.R
\name{scale_color_gradient_fg}
\alias{scale_color_gradient_fg}
\alias{scale_colour_gradient_fg}
\title{Wybor kontrastowego koloru}
\usage{
scale_color_gradient_fg(
  name = waiver(),
  ...,
  low = "#efefff",
  high = "#25304d",
  na.value = "grey50",
  guide = "none",
  fgToChoose = c("white", "black"),
  wcag = "3",
  aesthetics = "colour"
)

scale_colour_gradient_fg(
  name = waiver(),
  ...,
  low = "#efefff",
  high = "#25304d",
  na.value = "grey50",
  guide = "none",
  fgToChoose = c("white", "black"),
  wcag = "3",
  aesthetics = "colour"
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{...}{dodatkowe argumenty przekazywane do \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}}

\item{low, high}{Colours for low and high ends of the gradient.}

\item{na.value}{Colour to use for missing values}

\item{guide}{Type of legend. Use \code{"colourbar"} for continuous
colour bar, or \code{"legend"} for discrete colour legend.}

\item{fgToChoose}{wektor kolorów (opisanych w dowolny sposób zrozumiały dla
\code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}), spośród których wybierane będą bardziej kontrastowe
w stosunku do kolorów skali}

\item{wcag}{opcjonalnie wersja standardu WCAG (domyślnie 3), wg której ma być
obliczany kontrast (por. \link{compute_apca_contrast} i \link{compute_contrast_ratio})}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}
}
\description{
Umożliwia zdefiniowanie w wykresie \emph{\link[ggplot2:ggplot]{ggplot2}} skali kolorów
mających jak najwyższy kontrast w stosunku do skali wypełnienia zdefiniowanej
przy pomocy \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradient()}}.
}
\examples{
library(ggplot2)
}
\seealso{
\link{wykresMigracjeWoj}
}
