% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_fill_losy.R
\name{scale_color_losy}
\alias{scale_color_losy}
\alias{scale_colour_losy}
\title{Skale kolorow}
\usage{
scale_color_losy(
  name = waiver(),
  ...,
  type = c("statusy", "plec", "matura", "dyplomyZaw", "typySzkol", "kontynuacje"),
  aesthetics = "colour"
)

scale_colour_losy(
  name = waiver(),
  ...,
  type = c("statusy", "plec", "matura", "dyplomyZaw", "typySzkol", "kontynuacje"),
  aesthetics = "colour"
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{...}{dodatkowe argumenty przekazywane do \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}

\item{type}{ciąg znaków - typ skali}

\item{aesthetics}{The names of the aesthetics that this scale works with.}
}
\description{
Umożliwia zdefiniowanie w wykresie \emph{\link[ggplot2:ggplot]{ggplot2}} skali kolorów
spośród jednej z używanych w monitoringu karier absolwentów szkół
ponadpodstawowych.
}
\seealso{
\code{\link[=scale_fill_losy]{scale_fill_losy()}}, \code{\link[=scale_fill_losy_fg]{scale_fill_losy_fg()}}
}
