% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_fill_losy.R
\name{scale_fill_losy}
\alias{scale_fill_losy}
\title{Skale kolorow}
\usage{
scale_fill_losy(
  name = waiver(),
  ...,
  type = c("statusy", "plec", "matura", "dyplomyZaw", "typySzkol", "kontynuacje"),
  aesthetics = "fill"
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{...}{dodatkowe argumenty przekazywane do \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}

\item{type}{ciąg znaków - typ skali}

\item{aesthetics}{The names of the aesthetics that this scale works with.}
}
\description{
Umożliwia zdefiniowanie w wykresie \emph{\link[ggplot2:ggplot]{ggplot2}} skali kolorów
wypełnienia spośród jednej z używanych w monitoringu karier absolwentów szkół
ponadpodstawowych.
}
\examples{
library(ggplot2)
p <- ggplot(daneTypySzkol,
            aes(pct, factor(rok_abs), fill = typ_szk,
                label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE))
p
p + scale_fill_losy(type = "typySzkol")
}
\seealso{
\code{\link[=scale_color_losy]{scale_color_losy()}}, \code{\link[=scale_color_losy_fg]{scale_color_losy_fg()}},
\link{wykresDyplomy}, \link{wykresKontynuacjeDziedziny}, \link{wykresKontynuacjeTypSzk},
\link{wykresMatury}, \link{wykresPlec}, \link{wykresStatusy}, \link{wykresStatusyPlec},
\link{wykresTypySzkol}
}
