% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\docType{data}
\name{sfPow}
\alias{sfPow}
\title{Geometrie do kartogramow - powiaty}
\format{
Obiekt klasy \emph{sf}: ramka danych z 381 wierszami i 31 kolumnami.

Do łączenia z wynikami może zostać łatwo wykorzystana kolumna \code{teryt_pow},
która zawiera kody (3/4-cyfrowe) TERYT powiatów (oraz wiersz z brakiem
danych, przeznaczony do opisania, że powiat jest nieznany).

Zawiera też kolumnę \code{nazwa_pow} z nazwami powiatów (małymi literami, bez
poprzedzającego \emph{powiat}), ale akurat po niej (samej) łączyć danych \strong{nie}
należy, bo jej wartości nie są unikalne (jest w Polsce kilka par powiatów,
które nazywają się tak samo); powiat z brakiem danych w zmiennej \code{teryt_pow}
ma tu przypisaną wartość \emph{nieznany}.
}
\usage{
sfPow
}
\description{
Obiekt klasy \emph{sf} zawierający geometrie do rysowania kartogramów
na poziomie powiatów. Współrzędne zapisane są w układzie WGS 84, tzn. jako
stopnie długości i szerokości geograficznej.
}
\seealso{
\code{\link[=zlacz_z_mapa]{zlacz_z_mapa()}}, \link{sfWoj}
}
\keyword{datasets}
