% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_point_on_surface_sw.R
\name{st_point_on_surface_sw}
\alias{st_point_on_surface_sw}
\title{Przygotowanie danych do kartogramow}
\usage{
st_point_on_surface_sw(x)
}
\arguments{
\item{x}{obiekt klasy \emph{sfg}, \link[sf:sfc]{sf::sfc} lub \link[sf:sf]{sf::sf}}
}
\value{
obiekt tej samej klasy, co \code{x}, ale z przekształconą geometrią
(p. \code{\link[sf:geos_unary]{sf::st_point_on_surface()}} i \code{\link[sf:st_zm]{sf::st_zm()}})
}
\description{
Funkcja opakowuje wywołanie \code{sf::st_point_on_surface(sf::st_zm(x))}
w wywołanie \code{suppressWarnings()}, aby można było użyć jej jako argumentu
\code{fun.geometry} w wywołaniach \code{\link[ggplot2:ggsf]{ggplot2::geom_sf_text()}} lub
\code{\link[ggplot2:ggsf]{ggplot2::geom_sf_label()}}. Umożliwia to uniknięcie wywoływania przez,
używane w domyślny sposób, ww. funkcje pakietu \emph{ggplot2} ostrzeżeń o tym, że
\code{\link[sf:geos_unary]{sf::st_point_on_surface()}} może dawać nieprecyzyjne wyniki, jeśli jest
stosowana w odniesieniu do obiektów \link[sf:sf]{sf::sf} ze współrzędnymi opisanymi
w ramach projekcji sferycznych (tj. jako szerokość i długość geograficzna).
Wobec relatywnie niewielkiej rozległości Polski problem ten jest jednak
pomijalny (nie da się za to w inny sposób łatwo wyłączyć tego ostrzeżenia).

Nawiasem mówiąc, problem ten można by też rozwiązać używając geometrii
zapisanych w projekcji nie przysparzającej takich kłopotów (np. EPSG 2180, co
można uzyskać wywołując \code{sf::st_transform(sfWoj, 2180)}). Z tym że wiązało by
to się z pewną zmianą kształtu geometrii wyświetlanych na wykresie.
}
\examples{
st_point_on_surface_sw(sfWoj)
\dontrun{
  sf::st_point_on_surface(sfWoj) # generuje ostrzeżenie
}
}
\seealso{
\link{wykresMigracjeWoj}, \code{\link[ggplot2:ggsf]{ggplot2::geom_sf_text()}}
}
