% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_losy.R
\name{theme_losy}
\alias{theme_losy}
\title{Temat wykresu na potrzeby monitoringu karier}
\usage{
theme_losy(
  base_size = 18,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  axis_ticks = c("x", "y", "xy", "none")
)
}
\arguments{
\item{base_size}{liczba, p. \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}}

\item{base_family}{opcjonalnie ciąg znaków, p. \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}}

\item{base_line_size}{opcjonalnie liczba, p. \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}}

\item{base_rect_size}{opcjonalnie liczba, p. \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}}

\item{axis_ticks}{opcjonalnie ciąg znaków wskazujący, na której osi/osiach
wykresu mają zostać narysowane \emph{pręciki} przy podanych wartościach}
}
\description{
Nadaje wykresowi \emph{\link[ggplot2:ggtheme]{temat}} definiujący jego wygląd
(ale nie \emph{układ}) odpowiedni do wykorzystania na potrzeby systemu monitoringu
karier absolwentów szkół ponadpodstawowych.
}
\examples{
library(ggplot2)
p <- ggplot(daneTypySzkol,
            aes(pct, factor(rok_abs), fill = typ_szk,
                label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(aes(colour = typ_szk),
            position = position_stack(reverse = TRUE, vjust = 0.5),
            size = 7) +
  scale_x_continuous(NULL, labels = label_percent_pl(),
                     expand = c(0, 0, 0, 0.01)) +
  scale_y_discrete(NULL, expand = c(0, 0))  +
  scale_fill_losy(NULL, type = "typySzkol",
                  guide = guide_legend(ncol = 2)) +
  scale_colour_losy_fg(NULL, type = "typySzkol", guide = "none")
p
p +
  theme_losy(base_family = "sans")
}
\seealso{
\code{\link[=zmien_wielkosc_czcionek]{zmien_wielkosc_czcionek()}}, \code{\link[=zmien_rodzine_czcionek]{zmien_rodzine_czcionek()}},
\link{wykresBezrobocie}, \link{wykresBezrobocieOkresGrupaOdn}, \link{wykresBezrobociePlec},
\link{wykresBezrobocieZawod},
\link{wykresDyplomyPlec}, \link{wykresDyplomyZawod},
\link{wykresFormyZatrudnGrupaOdn}, \link{wykresKontMlodocPracEduGrupaOdn},
\link{wykresKontMlodocPracNeduGrupaOdn},
\link{wykresKontynuacjeDziedzinyPlec}, \link{wykresKontynuacjeDyscyplinyPlec},
\link{wykresKontynuacjeTypSzk}, \link{wykresKontynuacjeTypSzkPlec},
\link{wykresKontynuacjeTypSzkZawod},
\link{wykresMaturyPlec}, \link{wykresMaturyZawod},
\link{wykresPlec}, \link{wykresPracaOkresGrupaOdn}, \link{wykresPrzychodyGrupaOdn},
\link{wykresStatusy}, \link{wykresStatusyPlec}, \link{wykresStatusyZawod},
\link{wykresStatusyGrupaOdn}, \link{wykresTypySzkol}
}
