% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresBezrobocie}
\alias{wykresBezrobocie}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresBezrobocie
}
\description{
Wykres z przykładowym zestawieniem częstości bycia zarejestrowanym jako
bezrobotny ze względu na miesiąc od ukończenia szkoły.
}
\examples{
wykresBezrobocie

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneBezrobocie,
       aes(miesiac, bezrobocie, group = 1,
           label = percent_pl(bezrobocie, digits = 0))) +
  geom_line(linewidth = 0.7, colour = palette_losy("bezrobocie")["line"]) +
  geom_point(size = 5, colour = palette_losy("bezrobocie")["point"]) +
  geom_text(size = 6, nudge_y = -0.005) +
  scale_x_discrete("Miesiąc")  +
  scale_y_continuous("Odsetek bezrobotnych absolwentów",
                     labels = label_percent_pl(),
                     limits = c(0, NA_real_), expand = c(0, 0, 0.05, 0)) +
  theme_losy() +
  dodaj_rok_w_podpisie(rok = 3254)
}
\seealso{
\link{daneBezrobocie}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=palette_losy]{palette_losy()}}, \code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
