% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresBezrobocieOkresGrupaOdn}
\alias{wykresBezrobocieOkresGrupaOdn}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresBezrobocieOkresGrupaOdn
}
\description{
Wykres z przykładowym zestawieniem częstości bycia zarejestrowanym jako
bezrobotny pomiędzy daną szkołą a odpowiadającą jej \emph{grupą odniesienia} - do
wykorzystania w raportach szkolnych.
}
\examples{
wykresBezrobocieOkresGrupaOdn

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneBezrobocieOkresGrupaOdn,
       aes(okres_bezrobocia, pct, fill = grupa,
           label = percent_pl(pct, digits = 0))) +
  facet_wrap(vars(grupa), nrow = 1) +
  geom_col() +
  geom_text(size = 5, nudge_y = 0.03) +
  scale_x_discrete("Liczba miesięcy bezrobocia")  +
  scale_y_continuous("Odsetek absolwentów",
                     labels = label_percent_pl(),
                     limits = c(0, 1), expand = c(0, 0, 0.05, 0)) +
  scale_fill_losy(NULL, type = "grupaOdn", unname = TRUE, guide = "none") +
  theme_losy(axis_ticks = "y")
}
\seealso{
\link{daneBezrobocieOkresGrupaOdn}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
