% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresBezrobocieOkresGrupaOdnOld}
\alias{wykresBezrobocieOkresGrupaOdnOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresBezrobocieOkresGrupaOdnOld
}
\description{
Wykres z przykładowym zestawieniem częstości bycia zarejestrowanym jako
bezrobotny pomiędzy daną szkołą a odpowiadającą jej \emph{grupą odniesienia} - do
wykorzystania w raportach szkolnych.
}
\examples{
wykresBezrobocieOkresGrupaOdnOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresBezrobocieOkresGrupaOdn +
                   scale_fill_losy(NULL, type = "grupaOdn", unname = TRUE,
                                   variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresBezrobocieOkresGrupaOdn}
}
\keyword{datasets}
