% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresBezrobociePlecOld}
\alias{wykresBezrobociePlecOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresBezrobociePlecOld
}
\description{
Wykres z przykładowym zestawieniem częstości bycia
zarejestrowanym jako bezrobotny w podziale ze względu na płeć.
}
\examples{
wykresBezrobociePlecOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresBezrobociePlec +
                   scale_fill_losy(NULL, type = "bezrobocie",
                                   variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "bezrobocie",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresBezrobociePlec}
}
\keyword{datasets}
