% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresBezrobocieZawodOld}
\alias{wykresBezrobocieZawodOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresBezrobocieZawodOld
}
\description{
Wykres z przykładowym zestawieniem częstości bycia
zarejestrowanym jako bezrobotny  podziale ze względu na (najczęściej
wybierane) zawody wyuczone.
}
\examples{
wykresBezrobocieZawodOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresBezrobocieZawod +
                   scale_fill_losy(NULL, type = "bezrobocie",
                                   variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "bezrobocie",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresBezrobocieZawod}
}
\keyword{datasets}
