% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresDyplomyPlec}
\alias{wykresDyplomyPlec}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresDyplomyPlec
}
\description{
Wykres z przykładowym zestawieniem uzyskiwania
dyplomów w podziale ze względu na płeć.
}
\examples{
wykresDyplomyPlec

# kod tworzący ten wykres:
library(ggplot2)
}
\seealso{
\link{daneDyplomyPlec}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=scale_colour_losy_fg]{scale_colour_losy_fg()}}, \code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
