% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresDyplomyPlecOld}
\alias{wykresDyplomyPlecOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresDyplomyPlecOld
}
\description{
Wykres z przykładowym zestawieniem uzyskiwania
dyplomów w podziale ze względu na płeć.
}
\examples{
wykresDyplomyPlecOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresDyplomyPlec +
                   scale_fill_losy(NULL, type = "dyplomyZaw",
                                   variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "dyplomyZaw",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresDyplomyPlec}
}
\keyword{datasets}
