% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresDyplomyZawodOld}
\alias{wykresDyplomyZawodOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresDyplomyZawodOld
}
\description{
Wykres z przykładowym zestawieniem uzyskiwania
dyplomów w podziale ze względu na (najczęściej wybierane) zawody wyuczone.
}
\examples{
wykresDyplomyZawodOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresDyplomyZawod +
                   scale_fill_losy(NULL, type = "dyplomyZaw",
                                   variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "dyplomyZaw",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresDyplomyZawod}
}
\keyword{datasets}
