% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresFormyZatrudnGrupaOdn}
\alias{wykresFormyZatrudnGrupaOdn}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresFormyZatrudnGrupaOdn
}
\description{
Wykres z przykładowym zestawieniem formy zatrudnienia pomiędzy
daną szkołą a odpowiadającą jej \emph{grupą odniesienia} - do wykorzystania
w raportach szkolnych.
}
\examples{
wykresFormyZatrudnGrupaOdn

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneFormyZatrudnGrupaOdn,
       aes(grupa, pct, fill = forma_zatrudn,
           label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = FALSE)) +
  geom_text(aes(colour = forma_zatrudn),
            position = position_stack(reverse = FALSE, vjust = 0.5),
            size = 6) +
  scale_x_discrete(NULL, expand = c(0, 0)) +
  scale_y_continuous("Odsetek absolwentów", labels = label_percent_pl(),
                     expand = c(0, 0, 0, 0.01)) +
  scale_fill_losy(NULL, type = "formyZatrudn",
                  guide = guide_legend(ncol = 1)) +
  scale_colour_losy_fg(NULL, type = "formyZatrudn", guide = "none") +
  theme_losy(axis_ticks = "y")
}
\seealso{
\link{daneFormyZatrudnGrupaOdn}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=scale_colour_losy_fg]{scale_colour_losy_fg()}}, \code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
