% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresFormyZatrudnGrupaOdnOld}
\alias{wykresFormyZatrudnGrupaOdnOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresFormyZatrudnGrupaOdnOld
}
\description{
Wykres z przykładowym zestawieniem formy zatrudnienia pomiędzy
daną szkołą a odpowiadającą jej \emph{grupą odniesienia} - do wykorzystania
w raportach szkolnych.
}
\examples{
wykresFormyZatrudnGrupaOdnOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresFormyZatrudnGrupaOdn +
                   scale_fill_losy(NULL, type = "formyZatrudn",
                                   variantOld = TRUE,
                                   guide = guide_legend(ncol = 1)) +
                   scale_colour_losy_fg(NULL, type = "formyZatrudn",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresFormyZatrudnGrupaOdn}
}
\keyword{datasets}
