% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresKontMlodocPracEduGrupaOdn}
\alias{wykresKontMlodocPracEduGrupaOdn}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontMlodocPracEduGrupaOdn
}
\description{
Wykres z przykładowym zestawieniem częstości kontynuowania
pracy u tego samego pracodawcy, u którego absolwent był młodocianym
pracownikiem, wśród absolwentów kontynuujących naukę pomiędzy daną szkołą
a odpowiadającą jej \emph{grupą odniesienia} - do wykorzystania w raportach
szkolnych.
}
\examples{
wykresKontMlodocPracEduGrupaOdn

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneKontMlodocPracEduGrupaOdn,
       aes(pct, grupa, fill = kont_mlodoc_prac,
           label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(aes(colour = kont_mlodoc_prac),
            position = position_stack(reverse = TRUE, vjust = 0.5),
            size = 5) +
  scale_x_continuous(NULL, labels = label_percent_pl(),
                     expand = c(0, 0, 0, 0)) +
  scale_y_discrete(NULL, expand = c(0, 0)) +
  scale_fill_losy(NULL, type = "kontMlodocPracEdu",
                  guide = guide_legend(ncol = 1)) +
  scale_colour_losy_fg(NULL, type = "kontMlodocPracEdu", guide = "none") +
  theme_losy()
}
\seealso{
\link{daneKontMlodocPracEduGrupaOdn}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=scale_colour_losy_fg]{scale_colour_losy_fg()}}, \code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
