% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresKontMlodocPracEduGrupaOdnOld}
\alias{wykresKontMlodocPracEduGrupaOdnOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontMlodocPracEduGrupaOdnOld
}
\description{
Wykres z przykładowym zestawieniem częstości kontynuowania
pracy u tego samego pracodawcy, u którego absolwent był młodocianym
pracownikiem, wśród absolwentów kontynuujących naukę pomiędzy daną szkołą
a odpowiadającą jej \emph{grupą odniesienia} - do wykorzystania w raportach
szkolnych.
}
\examples{
wykresKontMlodocPracEduGrupaOdnOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresKontMlodocPracEduGrupaOdn +
                   scale_fill_losy(NULL, type = "kontMlodocPracEdu",
                                   variantOld = TRUE,
                                   guide = guide_legend(ncol = 1)) +
                   scale_colour_losy_fg(NULL, type = "kontMlodocPracEdu",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresKontMlodocPracEduGrupaOdn}
}
\keyword{datasets}
