% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresKontMlodocPracNeduGrupaOdnOld}
\alias{wykresKontMlodocPracNeduGrupaOdnOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontMlodocPracNeduGrupaOdnOld
}
\description{
Wykres z przykładowym zestawieniem częstości kontynuowania
pracy u tego samego pracodawcy, u którego absolwent był młodocianym
pracownikiem, wśród absolwentów niekontynuujących nauki pomiędzy daną szkołą
a odpowiadającą jej \emph{grupą odniesienia} - do wykorzystania w raportach
szkolnych.
}
\examples{
wykresKontMlodocPracNeduGrupaOdnOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresKontMlodocPracNeduGrupaOdn +
                   scale_fill_losy(NULL, type = "kontMlodocPracNedu",
                                   variantOld = TRUE,
                                   guide = guide_legend(ncol = 1)) +
                   scale_colour_losy_fg(NULL, type = "kontMlodocPracNedu",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresKontMlodocPracNeduGrupaOdn}
}
\keyword{datasets}
