% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresKontynuacjeDyscyplinyPlecOld}
\alias{wykresKontynuacjeDyscyplinyPlecOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontynuacjeDyscyplinyPlecOld
}
\description{
Wykres z przykładowym zestawieniem (najczęściej wybieranych)
dyscyplin kontynuowania nauki na studiach w podziale ze względu na płeć.
}
\examples{
wykresKontynuacjeDyscyplinyPlecOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresKontynuacjeDyscyplinyPlec +
                   scale_fill_losy(NULL, type = "plec", variantOld = TRUE,
                                   guide = guide_legend(ncol = 3)))
}
\seealso{
\link{wykresKontynuacjeDyscyplinyPlec}
}
\keyword{datasets}
