% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresKontynuacjeDziedzinyPlec}
\alias{wykresKontynuacjeDziedzinyPlec}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontynuacjeDziedzinyPlec
}
\description{
Wykres z przykładowym zestawieniem dziedzin
kontynuowania nauki na studiach w podziale ze względu na płeć.
}
\examples{
wykresKontynuacjeDziedzinyPlec

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneKontynuacjeDziedzinyPlec,
       aes(pct, dziedzina, fill = plec,
           label = percent_pl(pct, digits = 0))) +
  geom_col(position = position_dodge2(width = 1, reverse = TRUE)) +
  geom_text(position = position_dodge2(width = 1, reverse = TRUE),
            size = 5, hjust = -0.2) +
  scale_x_continuous(NULL, breaks = NULL, expand = c(0, 0, 0.2, 0)) +
  scale_y_discrete(NULL, expand = c(0, 0)) +
  scale_fill_losy(NULL, type = "plec",
                  guide = guide_legend(ncol = 3)) +
  theme_losy()  +
  dodaj_rok_w_podpisie(rok = 3254)
}
\seealso{
\link{daneKontynuacjeDziedzinyPlec}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
