% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresKontynuacjeDziedzinyPlecOld}
\alias{wykresKontynuacjeDziedzinyPlecOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontynuacjeDziedzinyPlecOld
}
\description{
Wykres z przykładowym zestawieniem dziedzin
kontynuowania nauki na studiach w podziale ze względu na płeć.
}
\examples{
wykresKontynuacjeDziedzinyPlecOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresKontynuacjeDziedzinyPlec +
                   scale_fill_losy(NULL, type = "plec", variantOld = TRUE,
                                   guide = guide_legend(ncol = 3)))
}
\seealso{
\link{wykresKontynuacjeDziedzinyPlec}
}
\keyword{datasets}
