% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresKontynuacjeTypSzk}
\alias{wykresKontynuacjeTypSzk}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontynuacjeTypSzk
}
\description{
Wykres z przykładowym zestawieniem częstości typu kontynuowania
nauki - do wykorzystania w raportach szkolnych.
}
\examples{
wykresKontynuacjeTypSzk

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneKontynuacjeTypSzk,
       aes(pct, kontynuacja, label = percent_pl(pct, digits = 0))) +
  geom_col(position = position_dodge2(width = 1, reverse = TRUE),
           fill = "#009899") +
  geom_text(position = position_dodge2(width = 1, reverse = TRUE),
            size = 6, hjust = -0.2) +
  scale_x_continuous(NULL, breaks = NULL, expand = c(0, 0, 0.2, 0)) +
  scale_y_discrete(NULL) +
  theme_losy()
}
\seealso{
\link{daneKontynuacjeTypSzk}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
