% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresKontynuacjeTypSzkOld}
\alias{wykresKontynuacjeTypSzkOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontynuacjeTypSzkOld
}
\description{
Wykres z przykładowym zestawieniem częstości typu kontynuowania
nauki - do wykorzystania w raportach szkolnych.
}
\examples{
wykresKontynuacjeTypSzkOld

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneKontynuacjeTypSzk,
       aes(pct, kontynuacja, label = percent_pl(pct, digits = 0))) +
  geom_col(position = position_dodge2(width = 1, reverse = TRUE),
           fill = "#90CF60") +
  geom_text(position = position_dodge2(width = 1, reverse = TRUE),
            size = 6, hjust = -0.2) +
  scale_x_continuous(NULL, breaks = NULL, expand = c(0, 0, 0.2, 0)) +
  scale_y_discrete(NULL) +
  theme_losy()
}
\seealso{
\link{wykresKontynuacjeTypSzk}
}
\keyword{datasets}
