% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresKontynuacjeTypSzkPlecOld}
\alias{wykresKontynuacjeTypSzkPlecOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontynuacjeTypSzkPlecOld
}
\description{
Wykres z przykładowym zestawieniem form kontynuowania
nauki w podziale ze względu na płeć.
}
\examples{
wykresKontynuacjeTypSzkPlecOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresKontynuacjeTypSzkPlec +
                   scale_fill_losy(NULL, type = "kontynuacje",
                                   variantOld = TRUE,
                                   guide = guide_legend(ncol = 3)))
}
\seealso{
\link{wykresKontynuacjeTypSzkPlec}
}
\keyword{datasets}
