% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresKontynuacjeTypSzkZawod}
\alias{wykresKontynuacjeTypSzkZawod}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresKontynuacjeTypSzkZawod
}
\description{
Wykres z przykładowym zestawieniem form kontynuowania
nauki w podziale ze względu na (najczęściej wybierane) zawody wyuczone.
}
\examples{
wykresKontynuacjeTypSzkZawod

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneKontynuacjeTypSzkZawod,
       aes(pct, nazwa_zaw, fill = kontynuacja,
           label = percent_pl(pct, digits = 0))) +
  geom_col(position = position_dodge2(width = 1, reverse = TRUE)) +
  geom_text(position = position_dodge2(width = 1, reverse = TRUE),
            size = 6, hjust = -0.2) +
  scale_x_continuous(NULL, breaks = NULL, expand = c(0, 0, 0.2, 0)) +
  scale_y_discrete(NULL) +
  scale_fill_losy(NULL, type = "kontynuacje",
                  guide = guide_legend(ncol = 3)) +
  theme_losy() +
  dodaj_rok_w_podpisie(rok = 3254)
}
\seealso{
\link{daneKontynuacjeTypSzkZawod}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
