% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresMaturyPlecOld}
\alias{wykresMaturyPlecOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresMaturyPlecOld
}
\description{
Wykres z przykładowym zestawieniem wyników matury w podziale
ze względu na płeć.
}
\examples{
wykresMaturyPlecOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresMaturyPlec +
                   scale_fill_losy(NULL, type = "matura", variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "matura",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresMaturyPlec}
}
\keyword{datasets}
