% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresMaturyZawodOld}
\alias{wykresMaturyZawodOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresMaturyZawodOld
}
\description{
Wykres z przykładowym zestawieniem wyników matury w podziale
ze względu na (najczęściej wybierane) zawody wyuczone
}
\examples{
wykresMaturyZawodOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresMaturyZawod +
                   scale_fill_losy(NULL, type = "matura", variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "matura",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresMaturyZawod}
}
\keyword{datasets}
