% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresMigracjeWoj}
\alias{wykresMigracjeWoj}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresMigracjeWoj
}
\description{
Wykres z przykładowym kartogramem ilustrującym migracje na
poziomie województw (procent absolwentów mieszkających w poszczególnych
województwach).
}
\examples{
wykresMigracjeWoj

# kod tworzący ten wykres:
library(ggplot2)
ggplot(zlacz_z_mapa(daneMigracjeWoj),
         aes(fill = pct, colour = pct,
             label = percent_pl(pct, digits = 1))) +
  geom_sf() +
  geom_sf_text(size = 5,
               fun.geometry = st_point_on_surface_sw) +
  geom_sf_text(aes(label = label), colour = "black", size = 5,
               hjust = "left", vjust = "bottom",
               data = function(x) {
                 x <- x[x$nazwa_woj == "nieznane", ]
                 x$label <- rep("Nieznane     \nwojewództwo", nrow(x))
                 if (nrow(x) > 0L) {
                   x$geometry <-
                     sf::st_sfc(sf::st_point(c(210000, 220000)),
                                crs = attributes(x$geometry)$crs)
                 }
                 return(x)
               },
               fun.geometry = st_point_on_surface_sw) +
  scale_x_continuous(NULL, labels = function(x) {return(rep("", length(x)))},
                     expand = c(0, 0)) +
  scale_y_continuous(NULL, labels = function(x) {return(rep("", length(x)))},
                     expand = c(0, 0)) +
  scale_fill_gradient("Procent absolwentów", low = "#efefff", high = "#25304d",
                      guide = "none") +
  scale_colour_gradient_fg(NULL, low = "#efefff", high = "#25304d") +
  theme_losy(axis_ticks = "none") +
  dodaj_rok_w_podpisie(rok = 3254)
}
\seealso{
\link{daneMigracjeWoj}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=theme_losy]{theme_losy()}},
\link{scale_colour_gradient_fg}, \code{\link[=st_point_on_surface_sw]{st_point_on_surface_sw()}}
}
\keyword{datasets}
