% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresPlec}
\alias{wykresPlec}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresPlec
}
\description{
Wykres z przykładowym zestawieniem płci w podziale ze względu
na typ ukończonej szkoły.
}
\examples{
wykresPlec

# kod tworzący ten wykres:
library(ggplot2)
ggplot(danePlec,
       aes(pct, typ_szk, fill = plec,
           label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(aes(colour = plec),
            position = position_stack(reverse = TRUE, vjust = 0.5),
            size = 7) +
  scale_x_continuous(NULL, labels = label_percent_pl(),
                     expand = c(0, 0, 0, 0.02)) +
  scale_y_discrete(NULL, expand = c(0, 0)) +
  scale_fill_losy(NULL, type = "plec",
                  guide = guide_legend(ncol = 2)) +
  scale_colour_losy_fg(NULL, type = "plec", guide = "none") +
  theme_losy() +
  dodaj_rok_w_podpisie(rok = 3254)
}
\seealso{
\link{danePlec}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=scale_colour_losy_fg]{scale_colour_losy_fg()}}, \code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
