% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresPlecOld}
\alias{wykresPlecOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresPlecOld
}
\description{
Wykres z przykładowym zestawieniem płci w podziale ze względu
na typ ukończonej szkoły.
}
\examples{
wykresPlecOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresPlec +
                   scale_fill_losy(NULL, type = "plec", variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "plec", variantOld = TRUE,
                                        guide = "none"))
}
\seealso{
\link{wykresPlec}
}
\keyword{datasets}
