% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresPracaOkresGrupaOdnOld}
\alias{wykresPracaOkresGrupaOdnOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresPracaOkresGrupaOdnOld
}
\description{
Wykres z przykładowym zestawieniem części okresu posiadania
pracy (w ramach okresu od września do grudnia roku ukończenia szkoły)
pomiędzy daną szkołą a odpowiadającą jej \emph{grupą odniesienia} - do
wykorzystania w raportach szkolnych.
}
\examples{
wykresPracaOkresGrupaOdnOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresPracaOkresGrupaOdn +
                   scale_fill_losy(NULL, type = "okresPracy",
                                   variantOld = TRUE,
                                   guide = guide_legend(nrow = 1)) +
                   scale_colour_losy_fg(NULL, type = "okresPracy",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresPracaOkresGrupaOdn}
}
\keyword{datasets}
