% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresPrzeplywyStatusyOld}
\alias{wykresPrzeplywyStatusyOld}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresPrzeplywyStatusyOld
}
\description{
Wykres z przykładowym zestawieniem częstości występowania
poszczególnych ścieżek edukacyjno-zawodowych, definiowanych jako kombinacje
statusu edukacyjno-zawodowego w trzech punktach czasu: w 6, 30 i 54 miesiące
od ukończenia szkoły.
}
\examples{
library(ggalluvial)
wykresPrzeplywyStatusyOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresPrzeplywyStatusy +
                   scale_fill_losy(NULL, type = "statusy",
                                   variantOld = TRUE) +
                   scale_colour_losy_fg(NULL, type = "statusy",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresPrzeplywyStatusy}
}
\keyword{datasets}
