% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresPrzychodyGrupaOdn}
\alias{wykresPrzychodyGrupaOdn}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresPrzychodyGrupaOdn
}
\description{
Wykres z przykładowym zestawieniem parametrów rozkładu
przychodów z pracy (etatowej) pomiędzy daną szkołą a odpowiadającą jej
\emph{grupą odniesienia} - do wykorzystania w raportach szkolnych.
}
\examples{
wykresPrzychodyGrupaOdn

# kod tworzący ten wykres:
library(ggplot2)
ggplot(danePrzychodyGrupaOdn,
       aes(grupa)) +
  geom_boxplot(aes(ymin = c05, lower = c25, middle = c50, upper = c75,
                   ymax = c95, fill = grupa),
               stat = "identity", staplewidth = 1, linewidth = 1,
               width = 0.6) +
  geom_text(aes(y = wartosc, label = percent_pl(wartosc, digits = 0)),
            data = ~data.frame(grupa = rep(.x$grupa, 5),
                               wartosc = unlist(.x[, c("c05", "c25", "c50",
                                                       "c75", "c95")])),
            size = 5, position = position_nudge(x = -0.4)) +
  scale_x_discrete(NULL, expand = c(0, 0.6, 0, 0.5)) +
  scale_y_continuous(paste(c("Przychody absolwentów z pracy etatowej",
                             "w odniesieniu do przeciętnego",
                             "wynagrodzenia w powiatach"), collapse = "\n"),
                     labels = label_percent_pl(),
                     limits = c(0, NA_real_), expand = c(0, 0, 0, 0.01)) +
  scale_fill_losy(NULL, type = "grupaOdn", unname = TRUE, guide = "none") +
  theme_losy(axis_ticks = "y")
}
\seealso{
\link{danePrzychodyGrupaOdn}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
