% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresPrzychodyGrupaOdnOld}
\alias{wykresPrzychodyGrupaOdnOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresPrzychodyGrupaOdnOld
}
\description{
Wykres z przykładowym zestawieniem parametrów rozkładu
przychodów z pracy (etatowej) pomiędzy daną szkołą a odpowiadającą jej
\emph{grupą odniesienia} - do wykorzystania w raportach szkolnych.
}
\examples{
wykresPrzychodyGrupaOdnOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresPrzychodyGrupaOdn +
                   scale_fill_losy(NULL, type = "grupaOdn", unname = TRUE,
                                   variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresPrzychodyGrupaOdn}
}
\keyword{datasets}
