% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresStatusy}
\alias{wykresStatusy}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresStatusy
}
\description{
Wykres z przykładowym zestawieniem statusów
edukacyjno-zawodowych ze względu na miesiąc od ukończenia szkoły.
}
\examples{
wykresStatusy

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneStatusy,
       aes(miesiac, pct, fill = status,
           label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(aes(colour = status),
            position = position_stack(reverse = TRUE, vjust = 0.5),
            size = 5) +
  scale_x_discrete("Miesiąc") +
  scale_y_continuous("Odsetek absolwentów", labels = label_percent_pl(),
                     expand = c(0, 0, 0, 0)) +
  scale_fill_losy(NULL, type = "statusy",
                  guide = guide_legend(ncol = 2)) +
  scale_colour_losy_fg(NULL, type = "statusy", guide = "none") +
  theme_losy() +
  dodaj_rok_w_podpisie(rok = 3254)
}
\seealso{
\link{daneStatusy}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=scale_colour_losy_fg]{scale_colour_losy_fg()}}, \code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
