% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{wykresStatusyGrupaOdn}
\alias{wykresStatusyGrupaOdn}
\title{Przykladowe wykresy}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresStatusyGrupaOdn
}
\description{
Wykres z przykładowym zestawieniem statusów
edukacyjno-zawodowych pomiędzy daną szkołą (ew. danym zawodem w danej szkole)
a odpowiadającą jej \emph{grupą odniesienia} - do wykorzystania w raportach
szkolnych.
}
\examples{
wykresStatusyGrupaOdn

# kod tworzący ten wykres:
library(ggplot2)
ggplot(daneStatusyGrupaOdn,
       aes(pct, grupa, fill = status,
           label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(aes(colour = status),
            position = position_stack(reverse = TRUE, vjust = 0.5),
            size = 7) +
  scale_x_continuous(NULL, labels = label_percent_pl(),
                     expand = c(0, 0, 0, 0.01)) +
  scale_y_discrete(NULL, expand = c(0, 0)) +
  scale_fill_losy(NULL, type = "statusy",
                  guide = guide_legend(ncol = 2)) +
  scale_colour_losy_fg(NULL, type = "statusy", guide = "none") +
  theme_losy()
}
\seealso{
\link{daneStatusyGrupaOdn}, \code{\link[=percent_pl]{percent_pl()}}, \code{\link[=scale_fill_losy]{scale_fill_losy()}},
\code{\link[=scale_colour_losy_fg]{scale_colour_losy_fg()}}, \code{\link[=theme_losy]{theme_losy()}}
}
\keyword{datasets}
