% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresStatusyGrupaOdnOld}
\alias{wykresStatusyGrupaOdnOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresStatusyGrupaOdnOld
}
\description{
Wykres z przykładowym zestawieniem statusów
edukacyjno-zawodowych pomiędzy daną szkołą (ew. danym zawodem w danej szkole)
a odpowiadającą jej \emph{grupą odniesienia} - do wykorzystania w raportach
szkolnych.
}
\examples{
wykresStatusyGrupaOdnOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresStatusyGrupaOdn +
                   scale_fill_losy(NULL, type = "statusy", variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "statusy",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresStatusyGrupaOdn}
}
\keyword{datasets}
