% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresStatusyZawodOld}
\alias{wykresStatusyZawodOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresStatusyZawodOld
}
\description{
Wykres z przykładowym zestawieniem statusów
edukacyjno-zawodowych ze względu na (najczęściej wybierane) zawody wyuczone.
}
\examples{
wykresStatusyZawodOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresStatusyZawod +
                   scale_fill_losy(NULL, type = "statusy", variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "statusy",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresStatusyZawod}
}
\keyword{datasets}
