% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresTypySzkolOld}
\alias{wykresTypySzkolOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresTypySzkolOld
}
\description{
Wykres z przykładowym zestawieniem typów ukończonych szkół.
}
\examples{
wykresTypySzkolOld

# kod tworzący ten wykres:
library(ggplot2)
suppressMessages(wykresTypySzkol +
                   scale_fill_losy(NULL, type = "typySzkol", variantOld = TRUE,
                                   guide = guide_legend(ncol = 2)) +
                   scale_colour_losy_fg(NULL, type = "typySzkol",
                                        variantOld = TRUE, guide = "none"))
}
\seealso{
\link{wykresTypySzkol}
}
\keyword{datasets}
