% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlacz_z_mapa.R
\name{zlacz_z_mapa}
\alias{zlacz_z_mapa}
\title{Przygotowanie danych do kartogramow}
\usage{
zlacz_z_mapa(dane, sf = "sfWoj", allSf = FALSE)
}
\arguments{
\item{dane}{ramka danych z danymi do połączenia; co do zasady powinna
zawierać kolumny \code{teryt_woj} (TERYT województwa w formacie 1/2-cyfrowym) lub
\code{nazwa_woj} (ciągi znaków z nazwami województw)}

\item{sf}{opcjonalnie ciąg znaków z nazwą obiektu \link[sf:sf]{sf::sf}, z którym dane
mają zostać połączone - obecnie jedyną akceptowaną wartością jest "sfWoj"}

\item{allSf}{opcjonalnie wartość logiczna - czy w połączonych danych mają
znaleźć się wszystkie województwa (w tym \emph{nieznane województwo}), nawet jeśli
niektóre z nich nie występują w danych przekazanych do połączenia?}
}
\value{
obiekt klasy \link[sf:sf]{sf::sf}
}
\description{
Funkcja przyłącza dane opisujące wskaźniki obliczone na poziomie województw
do obiektu klasy \link[sf:sf]{sf::sf} zawierającego mapę, która ma zostać użyta do
narysowania kartogramu.
}
\details{
Obiekt \link{sfWoj} zawiera również rekord opisujący \emph{nieznane} województwo - jest
on opisywany przez wartość kolumny \code{taryt_woj} równą \code{NA_integer_} i wartość
kolumny \code{nazwa_woj} równą \code{"nieznane"}. Można w ten sposób zwizualizować na
kartogramie również wartość wskaźnika odnoszącą się do tych absolwentów, dla
których nie jest znana ich lokalizacja (miejsce zamieszkania, studiowania
itp.).
}
\examples{
zlacz_z_mapa(daneMigracjeWoj)
}
\seealso{
\link{wykresMigracjeWoj}
}
