% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmien_prog_pokazywania_etykiet.R
\name{zmien_prog_pokazywania_etykiet}
\alias{zmien_prog_pokazywania_etykiet}
\title{Modyfikowanie elementow wykresu}
\usage{
zmien_prog_pokazywania_etykiet(g, hideBelow = NULL, label = NULL)
}
\arguments{
\item{g}{wykres \emph{\link[ggplot2:ggplot]{ggplot2}}}

\item{hideBelow}{opcjonalnie wartość liczbowa: wartości etykiet (\emph{label})
poniżej tego progu będą na wykresie ukrywane (p. opis szczegółowy poniżej)}

\item{label}{opcjonalnie wyrażenie (\emph{expression}), które zostanie przypisane
do \emph{mapowań} \code{label} we wszystkich \emph{warstwach} wykresu (p. opis szczegółowy
poniżej) \strong{otoczone wywołaniem funkcji \code{\link[=quote]{quote()}}} (p. przykłady użycia)}
}
\value{
wykres \emph{\link[ggplot2:ggplot]{ggplot2}}
}
\description{
Zmienia próg wartości, poniżej którego ukrywane są na wykresie etykiety
wartości (\emph{label}).
}
\details{
W wywołaniu funkcji można podać wartość tylko argumentu \code{hideBelow} albo
tylko argumentu \code{label}.

Wywołania z podanym argumentem \code{hideBelow} \strong{ma sens tylko wtedy}, kiedy
przy tworzeniu modyfikowanego właśnie wykresu do zdefiniowania \emph{mapowań}
\emph{estetyki} \code{label} zostały użyte wywołania funkcji \code{\link[=percent_pl]{percent_pl()}} (lub ew.
\code{\link[=label_percent_pl]{label_percent_pl()}}) - w tym trybie funkcja poszukuje argumentów \code{hideBelow}
w wyrażeniach przypisanych do \emph{estetyki} \code{label} i zmienia przekazywaną nimi
wartość na podaną, ale jeśli takich argumentów nie znajdzie, to niczego nie
zmieni! (istnieją też inne ograniczenia - p. \code{\link[=zmien_hideBelow]{zmien_hideBelow()}})

Wywołanie z podanym argumentem \code{label} jest bardziej elastyczne i pozwala
przypisać do \emph{mapowania} \emph{estetyki} \code{label} całe podane wyrażenie.
Ograniczenie stanowi tu z jednej strony to, że będzie ono takie samo dla
wszystkich takich przypisań (co jednak w kontekście monitoringu karier co do
zasady nie powinno być problemem), a z drugiej strony, że użytkownik sam musi
się upewnić, czy podawana w wyrażeniu zmienna istnieje w danych danego
wykresu. W celu osiągnięcia efektu ukrycia etykiet poniżej pewnego progu
typowo argumentem \code{label} podawane będzie wyrażenie typu:
\code{percent_pl(pct, digits = pozadanaDokladnosc, hideBelow = pozadanyProg)},
w którym zamiast \code{pozadanaDokladnosc} i \code{pozadanyProg} wpisane zostaną
konkretne liczby.
}
\examples{
library(ggplot2)
p <- ggplot(daneTypySzkol,
            aes(pct, factor(rok_abs), fill = typ_szk,
                label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(position = position_stack(reverse = TRUE, vjust = 0.5),
            colour = "black")

p
zmien_prog_pokazywania_etykiet(p, hideBelow = 0.1)
# argument `label` daje nieco więcej elastyczności,
# ale wymaga otoczenie podanego wyrażenia wywołaniem funkcji `quote()`:
zmien_prog_pokazywania_etykiet(p, label = quote(percent_pl(pct, digits = 1,
                                                           hideBelow = 0.1)))
# uwaga, korzystając z argumentu `label` można też całkiem zmienić
# przypisanie do estetyk label:
zmien_prog_pokazywania_etykiet(p, label = quote(typ_szk))
}
\seealso{
\code{\link[=percent_pl]{percent_pl()}}, \code{\link[=zmien_wielkosc_czcionek]{zmien_wielkosc_czcionek()}},
\code{\link[=zmien_wielkosc_etykiet]{zmien_wielkosc_etykiet()}}, \code{\link[=zmien_rodzine_czcionek]{zmien_rodzine_czcionek()}}
}
