% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmien_wielkosc_czcionek.R
\name{zmien_wielkosc_czcionek}
\alias{zmien_wielkosc_czcionek}
\alias{zmien_wielkosc_etykiet}
\title{Modyfikowanie elementow wykresu}
\usage{
zmien_wielkosc_czcionek(g, baseSize = NULL, rescale = NULL, geoms = TRUE)

zmien_wielkosc_etykiet(g, size = NULL, rescale = NULL)
}
\arguments{
\item{g}{wykres \emph{\link[ggplot2:ggplot]{ggplot2}}}

\item{baseSize}{opcjonalnie liczba - bazowa wielkość czcionki - działa jak
argument \code{base_size} w funkcjach pakietu \emph{ggplot2} ustawiających
\emph{\link[ggplot2:ggtheme]{predefiniowane tematy}} wykresu (por. szczegółowy
opis poniżej)}

\item{rescale}{opcjonalnie liczba, przez którą przemnożona zostanie wielkość
wszystkich czcionek w elementach tekstowych \emph{\link[ggplot2:theme]{tematu}}
wykresu \emph{ggplot2} oraz/albo wielkość czcionek wszystkich \emph{geometrii}
tekstowych wykresu}

\item{geoms}{wartość logiczna (domyślnie \code{TRUE}), wskazująca, czy
zmieniona ma być również wielkość czcionek w \emph{geometriach} (zarówno
zdefiniowanych na wykresie, jak i w domyślnej \emph{geometrii} w \emph{temacie} danego
wykresu)}

\item{size}{opcjonalnie liczba - wielkość czcionki (w mm), jaka zostanie
nadana wszystkim \emph{geometriom} tekstowym wykresu}
}
\value{
wykres \emph{\link[ggplot2:ggplot]{ggplot2}}
}
\description{
Funkcje zmieniające wielkość elementów tekstowych na wykresie
\emph{\link[ggplot2:ggplot]{ggplot2}}:
\itemize{
\item \code{zmien_wielkosc_czcionek()} pozwala zmienić jednocześnie wielkość
elementów tekstowych \emph{tematu} wykresu (etykiety i wartości osi, elementy
legendy, tytuły itp.) oraz wielkość \emph{geometrii} (tekstowych),
przy czym zmianę wielkości tych drugich można zablokować korzystając
z argumentu \code{geoms=FALSE},
\item \code{zmien_wielkosc_etykiet()} pozwala zmienić wyłącznie wielkość \emph{geometrii}
(tekstowych).
}

Jako \emph{geometrie} tekstowe traktowane są te, które zostały utworzone za pomocą
funkcji \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} lub \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}. W praktyce
monitoringu karier odpowiadają one za prezentowanie etykiet wartości
procentowych na wykresach słupkowych.

Wielkość \emph{geometrii} tekstowych zostanie zmieniona \strong{tylko wtedy}, jeśli
\strong{nie} jest \emph{mapowana} z danych (w wywołaniu \code{\link[ggplot2:aes]{ggplot2::aes()}} dla całego
wykresu lub danej \emph{geometrii}), tzn. jest stała dla wszystkich jednostek
prezentowanych daną \emph{geometrią}.

Jeśli \code{geoms=TRUE}, za pośrednictwem zmian w \emph{temacie} wykresu
**we wszystkich \emph{warstwach} ** zostaną też zmienione wartości wszystkich
\emph{estetyk}, które mapują się na domyślną wartość \emph{estetyki} \code{fontsize}
definiowaną w ramach elementu \code{geom} \emph{tematu} danego wykresu
(por. \code{\link[ggplot2:aes_eval]{ggplot2::from_theme()}}).
}
\details{
W wywołaniach funkcji można podać wartość tylko argumentu \code{rescale} albo
tylko argumentu \code{baseSize}/\code{size}.

Zmiana wielkości przez \code{zmien_wielkosc_czcionek()} jest zaimplementowana
w ten sposób, że jeśli podany został argument \code{baseSize}, to jego wartość
jest dzielona przez dotychczasową wartość wielkości elementu \code{text} \emph{tematu}
danego wykresu, a uzyskana wartość używana jest następnie tak, jakby została
podana przy pomocy argumentu \code{rescale}. Funkcja przechodzi przez wszystkie
elementy tekstowe \emph{tematu} danego wykres i jeśli wielkość jest w nich
ustawiona w sposób absolutny, modyfikuje ją mnożąc przez wartość argumentu
\code{rescale} (jeśli wielkość jakiegoś elementu tekstowego jest zdefiniowana
relatywnie, jej definicja zostaje pozostawiona bez zmian, gdyż zmiana
wielkości zostanie na nią spropagowana automatycznie). Jeśli dany wykres nie
miał jeszcze przypisanego żadnego \emph{tematu}, zostanie mu przypisany temat
zwrócony przez \code{\link[ggplot2:get_theme]{ggplot2::theme_get()}}, a następnie dokonane zostaną
modyfikacje wielkości elementów tekstowych.
}
\examples{
library(ggplot2)
p <- ggplot(daneTypySzkol,
            aes(pct, factor(rok_abs), fill = typ_szk,
                label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(position = position_stack(reverse = TRUE, vjust = 0.5),
            colour = "black")

p
zmien_wielkosc_czcionek(p, baseSize = 20)
zmien_wielkosc_czcionek(p, baseSize = 20, geoms = FALSE)
zmien_wielkosc_czcionek(p, rescale = 1.5)
zmien_wielkosc_czcionek(p, rescale = 1.5, geoms = FALSE)
# lub w pipie:
p |> zmien_wielkosc_czcionek(baseSize = 20)

zmien_wielkosc_etykiet(p, size = 10)
p |> zmien_wielkosc_etykiet(rescale = 0.8)

# jeśli wielkość geometri jest mapowana z danych, nie zostanie zmieniona:
ps <- ggplot(daneTypySzkol,
             aes(pct, factor(rok_abs), fill = typ_szk,
             label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(aes(size = pct),
            position = position_stack(reverse = TRUE, vjust = 0.5),
            colour = "black")
ps
zmien_wielkosc_etykiet(ps, size = 10)
zmien_wielkosc_czcionek(ps, baseSize = 20)
}
\seealso{
\code{\link[=zmien_rodzine_czcionek]{zmien_rodzine_czcionek()}}, \code{\link[=zmien_prog_pokazywania_etykiet]{zmien_prog_pokazywania_etykiet()}}
}
