% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette_fun.R
\name{create_palette_losy_fun}
\alias{create_palette_losy_fun}
\title{Skale kolorow}
\usage{
create_palette_losy_fun(type, unname = FALSE, variantOld = FALSE)
}
\arguments{
\item{type}{ciąg znaków - typ skali}

\item{unname}{wartość logiczna - czy zwracany wektor ma mieć przypisane
nazwy (opisujące, jakiej wartości ma odpowiadać dany kolor)?
(konieczne, kiedy mapowana zmienna nie ma stałego zestawu wartości, jak
w przypadku raportów szkolnych, gdy kolorem oznaczana jest dana szkoła
w odróżnieniu od jej \emph{grupy odniesienia}, p. \link{wykresBezrobocieOkresGrupaOdn}
i \link{wykresPrzychodyGrupaOdn})}

\item{variantOld}{wartość logiczna - czy zwrócić palety używane w raportach
do 2024 r.?}
}
\value{
Funkcja zwracająca paletę kolorów (przyjmująca jako parametr liczbę
kolorów w zwracanej palecie, która jednak będzie działać tylko dla jednej
liczby kolorów - kompatybilnej z liczbą kolorów w skali kolorów podanego
typu, zwracanej przez \code{\link[=palette_losy]{palette_losy()}}).
}
\description{
Nieeksportowana fabryka funkcji do wykorzystania wewnętrznie przez
\code{\link[=scale_color_losy]{scale_color_losy()}} i \code{\link[=scale_fill_losy]{scale_fill_losy()}} w ramach mechanizmu skal kolorów
\emph{ggplot2}.
}
