% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daneKontynuacjeTypSzk}
\alias{daneKontynuacjeTypSzk}
\title{Przykladowe dane}
\format{
Ramka danych z 5 wierszami i 2 kolumnami:
\itemize{
\item{\code{kontynuacja} - czynnik o poziomach \emph{Nie kontynuują kształcenia}, \emph{Kurs umiejętności zawodowych}, \emph{Kwalifikacyjny kurs zawodowy}, \emph{Szkoła policealna}, \emph{Szkoła wyższa} (w zależności od typu szkoły może to też być: \emph{Nie kontynuują kształcenia}, \emph{Kurs umiejętności zawodowych}, \emph{Kwalifikacyjny kurs zawodowy}, \emph{Branżowa szkoła II stopnia}, \emph{Liceum dla dorosłych})}
\item{\code{pct} - częstość wyboru danego typu kształcenia.}
}
}
\usage{
daneKontynuacjeTypSzk
}
\description{
Przykładowe zestawienie częstości typu kontynuowania nauki - do
wykorzystania w raportach szkolnych.
}
\seealso{
\link{wykresKontynuacjeTypSzkPlec}
}
\keyword{datasets}
