% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_losy.R
\name{palette_losy}
\alias{palette_losy}
\title{Skale kolorow}
\usage{
palette_losy(
  type = c("bezrobocie", "dyplomyZaw", "formyZatrudn", "grupaOdn", "kontMlodocPracEdu",
    "kontMlodocPracNedu", "kontynuacje", "matura", "okresPracy", "plec", "statusy",
    "typySzkol"),
  unname = FALSE,
  variantOld = FALSE
)
}
\arguments{
\item{type}{ciąg znaków - typ skali}

\item{unname}{wartość logiczna - czy zwracany wektor ma mieć przypisane
nazwy (opisujące, jakiej wartości ma odpowiadać dany kolor)?
(konieczne, kiedy mapowana zmienna nie ma stałego zestawu wartości, jak
w przypadku raportów szkolnych, gdy kolorem oznaczana jest dana szkoła
w odróżnieniu od jej \emph{grupy odniesienia}, p. \link{wykresBezrobocieOkresGrupaOdn}
i \link{wykresPrzychodyGrupaOdn})}

\item{variantOld}{wartość logiczna - czy zwrócić palety używane w raportach
do 2024 r.?}
}
\value{
wektor ciągów znaków z kodami kolorów palety
}
\description{
Zwraca skale kolorów wykorzystywane w monitoringu karier absolwentów szkół
ponadpodstawowych.#'
}
