% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_fill_losy.R
\name{scale_color_losy}
\alias{scale_color_losy}
\alias{scale_colour_losy}
\title{Skale kolorow}
\usage{
scale_color_losy(
  name = waiver(),
  ...,
  type = eval(formals(LOSYkolory::palette_losy)$type),
  unname = FALSE,
  aesthetics = "colour"
)

scale_colour_losy(
  name = waiver(),
  ...,
  type = eval(formals(LOSYkolory::palette_losy)$type),
  unname = FALSE,
  aesthetics = "colour"
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{...}{dodatkowe argumenty przekazywane do \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}

\item{type}{ciąg znaków - typ skali}

\item{unname}{wartość logiczna - czy zwracany wektor ma mieć przypisane
nazwy (opisujące, jakiej wartości ma odpowiadać dany kolor)?
(konieczne, kiedy mapowana zmienna nie ma stałego zestawu wartości, jak
w przypadku raportów szkolnych, gdy kolorem oznaczana jest dana szkoła
w odróżnieniu od jej \emph{grupy odniesienia}, p. \link{wykresBezrobocieOkresGrupaOdn}
i \link{wykresPrzychodyGrupaOdn})}

\item{aesthetics}{The names of the aesthetics that this scale works with.}
}
\description{
Umożliwia zdefiniowanie w wykresie \emph{\link[ggplot2:ggplot]{ggplot2}} skali kolorów
spośród jednej z używanych w monitoringu karier absolwentów szkół
ponadpodstawowych.
}
\seealso{
\code{\link[=scale_fill_losy]{scale_fill_losy()}}, \code{\link[=scale_fill_losy_fg]{scale_fill_losy_fg()}}
}
