% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_fill_losy_fg.R
\name{scale_color_losy_fg}
\alias{scale_color_losy_fg}
\alias{scale_colour_losy_fg}
\title{Wybor kontrastowego koloru}
\usage{
scale_color_losy_fg(
  name = waiver(),
  ...,
  type = eval(formals(LOSYkolory::palette_losy)$type),
  fgToChoose = c("white", "black"),
  wcag = "3",
  unname = FALSE,
  variantOld = FALSE,
  aesthetics = "colour"
)

scale_colour_losy_fg(
  name = waiver(),
  ...,
  type = eval(formals(LOSYkolory::palette_losy)$type),
  fgToChoose = c("white", "black"),
  wcag = "3",
  unname = FALSE,
  variantOld = FALSE,
  aesthetics = "colour"
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{...}{dodatkowe argumenty przekazywane do \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}

\item{type}{ciąg znaków - typ skali}

\item{fgToChoose}{wektor kolorów (opisanych w dowolny sposób zrozumiały dla
\code{\link[grDevices:col2rgb]{grDevices::col2rgb()}}), spośród których wybierane będą bardziej kontrastowe
w stosunku do kolorów skali, na którą wskazuje wartość argumentu \code{type}}

\item{wcag}{opcjonalnie wersja standardu WCAG (domyślnie 3), wg której ma być
obliczany kontrast (por. \link{compute_apca_contrast} i \link{compute_contrast_ratio})}

\item{unname}{wartość logiczna - czy zwracany wektor ma mieć przypisane
nazwy (opisujące, jakiej wartości ma odpowiadać dany kolor)?
(konieczne, kiedy mapowana zmienna nie ma stałego zestawu wartości, jak
w przypadku raportów szkolnych, gdy kolorem oznaczana jest dana szkoła
w odróżnieniu od jej \emph{grupy odniesienia}, p. \link{wykresBezrobocieOkresGrupaOdn}
i \link{wykresPrzychodyGrupaOdn})}

\item{variantOld}{wartość logiczna - czy zwrócić palety używane w raportach
do 2024 r.?}

\item{aesthetics}{The names of the aesthetics that this scale works with.}
}
\description{
Umożliwia zdefiniowanie w wykresie \emph{\link[ggplot2:ggplot]{ggplot2}} skali kolorów
mających jak najwyższy kontrast w stosunku do odpowiedniej ze zdefiniowanych
skal kolorów używanych w monitoringu karier absolwentów szkół
ponadpodstawowych.
}
\examples{
library(ggplot2)
p <- ggplot(daneTypySzkol,
            aes(pct, factor(rok_abs), fill = typ_szk,
                label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(aes(colour = typ_szk),
            position = position_stack(reverse = TRUE, vjust = 0.5),
            size = 7) +
  scale_fill_losy(type = "typySzkol")
p
p + scale_color_losy_fg(type = "typySzkol", guide = "none")
}
\seealso{
\code{\link[=scale_fill_losy]{scale_fill_losy()}}, \code{\link[=pick_more_contrasting]{pick_more_contrasting()}}, \code{\link[=scale_fill_losy_fg]{scale_fill_losy_fg()}},
\link{wykresBezrobociePlec}, \link{wykresBezrobocieZawod},
\link{wykresDyplomyPlec}, \link{wykresDyplomyZawod},
\link{wykresFormyZatrudnGrupaOdn}, \link{wykresKontMlodocPracEduGrupaOdn},
\link{wykresKontMlodocPracNeduGrupaOdn},
\link{wykresMaturyPlec}, \link{wykresMaturyZawod},
\link{wykresPlec}, \link{wykresPracaOkresGrupaOdn},
\link{wykresStatusy}, \link{wykresStatusyPlec}, \link{wykresStatusyZawod},
\link{wykresStatusyGrupaOdn}, \link{wykresTypySzkol}
}
