% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_fill_losy.R
\name{scale_fill_losy}
\alias{scale_fill_losy}
\title{Skale kolorow}
\usage{
scale_fill_losy(
  name = waiver(),
  ...,
  type = eval(formals(LOSYkolory::palette_losy)$type),
  unname = FALSE,
  aesthetics = "fill"
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{...}{dodatkowe argumenty przekazywane do \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}

\item{type}{ciąg znaków - typ skali}

\item{unname}{wartość logiczna - czy zwracany wektor ma mieć przypisane
nazwy (opisujące, jakiej wartości ma odpowiadać dany kolor)?
(konieczne, kiedy mapowana zmienna nie ma stałego zestawu wartości, jak
w przypadku raportów szkolnych, gdy kolorem oznaczana jest dana szkoła
w odróżnieniu od jej \emph{grupy odniesienia}, p. \link{wykresBezrobocieOkresGrupaOdn}
i \link{wykresPrzychodyGrupaOdn})}

\item{aesthetics}{The names of the aesthetics that this scale works with.}
}
\description{
Umożliwia zdefiniowanie w wykresie \emph{\link[ggplot2:ggplot]{ggplot2}} skali kolorów
wypełnienia spośród jednej z używanych w monitoringu karier absolwentów szkół
ponadpodstawowych.
}
\examples{
library(ggplot2)
p <- ggplot(daneTypySzkol,
            aes(pct, factor(rok_abs), fill = typ_szk,
                label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE))
p
p + scale_fill_losy(type = "typySzkol")
}
\seealso{
\code{\link[=scale_color_losy]{scale_color_losy()}}, \code{\link[=scale_color_losy_fg]{scale_color_losy_fg()}},
\link{wykresBezrobocieOkresGrupaOdn}, \link{wykresBezrobociePlec},
\link{wykresBezrobocieZawod},
\link{wykresDyplomyPlec}, \link{wykresDyplomyZawod},
\link{wykresFormyZatrudnGrupaOdn}, \link{wykresKontMlodocPracEduGrupaOdn},
\link{wykresKontMlodocPracNeduGrupaOdn},
\link{wykresKontynuacjeDziedzinyPlec}, \link{wykresKontynuacjeDyscyplinyPlec},
\link{wykresKontynuacjeTypSzkPlec}, \link{wykresKontynuacjeTypSzkZawod},
\link{wykresMaturyPlec}, \link{wykresMaturyZawod},
\link{wykresPlec}, \link{wykresPracaOkresGrupaOdn}, \link{wykresPrzychodyGrupaOdn},
\link{wykresStatusy}, \link{wykresStatusyPlec}, \link{wykresStatusyZawod},
\link{wykresStatusyGrupaOdn}, \link{wykresTypySzkol}
}
