% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs.R
\docType{data}
\name{sfWoj}
\alias{sfWoj}
\title{Geometrie do kartogramow - wojewodztwa}
\format{
Obiekt klasy \emph{sf}: ramka danych z 20 wierszami i 37 kolumnami. Do
łączenia z wynikami mogą zostać łatwo wykorzystane kolumny:
\itemize{
\item{\code{teryt_woj} - kody (1/2-cyforwe) TERYT województw; zawiera też wiersz z brakiem danych, przeznaczony do opisania, że województo jest nieznane,}
\item{\code{nazwa_woj} - nazwy województw (małymi literami); zawiera też dodatkowo województwo opisane jako \emph{nieznane}.}
}
}
\usage{
sfWoj
}
\description{
Obiekt klasy \emph{sf} zawierający geometrie do rysowania kartogramów
na poziomie województw.
}
\seealso{
\link{daneMigracjeWoj}, \code{\link[=zlacz_z_mapa]{zlacz_z_mapa()}}, \link{wykresMigracjeWoj}
}
\keyword{datasets}
