% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsOld.R
\docType{data}
\name{wykresBezrobocieOld}
\alias{wykresBezrobocieOld}
\title{Przykladowe wykresy - stare palety kolorow}
\format{
Wykres \emph{\link[ggplot2:ggplot]{ggplot2}}.
}
\usage{
wykresBezrobocieOld
}
\description{
Wykres z przykładowym zestawieniem częstości bycia zarejestrowanym jako
bezrobotny ze względu na miesiąc od ukończenia szkoły.
}
\examples{
wykresBezrobocieOld

# kod tworzący ten wykres:
library(ggplot2)
# wykres jest tożsamy z tym w nowej skali kolorów
wykresBezrobocie
}
\seealso{
\link{wykresBezrobocie}
}
\keyword{datasets}
