% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmien_rodzine_czcionek.R
\name{zmien_rodzine_czcionek}
\alias{zmien_rodzine_czcionek}
\title{Modyfikowanie elementow wykresu}
\usage{
zmien_rodzine_czcionek(g, family)
}
\arguments{
\item{g}{wykres \emph{\link[ggplot2:ggplot]{ggplot2}}}

\item{family}{ciąg znaków - nazwa rodziny czcionek, która
zostanie nadana wszystkim \emph{geometriom} tekstowym; można też przekazać brak
danych (\code{NA_character}), aby nakazać skopiowanie rodziny czcionek z ustawień
\emph{tematu} danego wykresu (dokładnie, z elementu \code{text})}
}
\value{
wykres \emph{\link[ggplot2:ggplot]{ggplot2}}
}
\description{
Zmienia rodzinę czcionek używanych do wyświetlania \emph{geometrii} tekstowych
na wykresie \emph{\link[ggplot2:ggplot]{ggplot2}}.

\strong{Wielkość} \emph{geometrii} tekstowych zostanie zmieniona \strong{tylko wtedy}, jeśli
\strong{nie} jest \emph{mapowana} z danych (w wywołaniu \code{\link[ggplot2:aes]{ggplot2::aes()}} dla całego
wykresu lub danej \emph{geometrii}), tzn. jest stała dla wszystkich jednostek
prezentowanych daną \emph{geometrią}.
}
\details{
Jeśli dany wykres nie ma jeszcze przypisanego żadnego \emph{tematu}, a jako
argument \code{family} podano brak danych, zostanie użyta rodzina czcionek
\emph{tematu} zwróconego przez \code{\link[ggplot2:theme_get]{ggplot2::theme_get()}} (choć skądinąd może to być
pusty ciąg znaków).
}
\examples{
library(ggplot2)
p <- ggplot(daneTypySzkol,
            aes(pct, factor(rok_abs), fill = typ_szk,
                label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(position = position_stack(reverse = TRUE, vjust = 0.5),
            size = 8, colour = "black")

p
zmien_rodzine_czcionek(p, "serif")
# lub w pipie:
p |> zmien_rodzine_czcionek("serif")

# jeśli rodzina czcionek geometri jest mapowana z danych, nie zostanie zmieniona:
ps <- ggplot(cbind(daneTypySzkol, czcionka = "sans"),
             aes(pct, factor(rok_abs), fill = typ_szk,
             label = percent_pl(pct, digits = 0, hideBelow = 0.05))) +
  geom_col(position = position_stack(reverse = TRUE)) +
  geom_text(aes(family = czcionka),
            position = position_stack(reverse = TRUE, vjust = 0.5),
            size = 8, colour = "black")
ps
zmien_rodzine_czcionek(ps, "serif")
}
\seealso{
\code{\link[=zmien_wielkosc_czcionek]{zmien_wielkosc_czcionek()}}, \code{\link[=zmien_wielkosc_etykiet]{zmien_wielkosc_etykiet()}},
\code{\link[=zmien_prog_pokazywania_etykiet]{zmien_prog_pokazywania_etykiet()}}
}
